Example energy conversion chain to demonstrate perfect substitution

A dataset containing an example energy conversion chain for the purpose of demonstrating changes to industry inputs where the inputs are perfect substitutes.

PerfectSubtidy

Format

A data frame with 20 rows and 10 variables:

Country

country, (Example, only one country)

Year

year, 2000 (only one year as an example)

Ledger.side

Supply or Consumption

Flow.aggregation.point

tells where each row should be aggregated

Energy.type

E.ktoe (for energy) or X.ktoe (for exergy)

Last.stage

tells the final stage of the energy conversion chain: services is the only entry here

Flow

the Industry or Sector involved in this flow

Product

the energy product involved in this flow

EX.ktoe

value of the energy, exergy, or service flow in ktoe

Unit

unit in which quantity is expressed

Details

PerfectSubtidy gives each non-zero entry in PerfectSubmats as a single column in a data frame. These data are in tidy format.