Add a column of matrix names to tidy data frame

This function adds a column of matrix names to a tidy data frame wherein each row of .DF is a single value in an energy conversion chain. The default argument values assume that .DF uses IEA-style nomenclature and terminology, although .DF does not necessarily need to contain IEA data.

add_matnames_iea(.DF, ledger_side = "Ledger.side", energy = "EX.ktoe",
  flow_aggregation_point = "Flow.aggregation.point", flow = "Flow",
  supply_side = "Supply", consumption_side = "Consumption",
  eiou = "Energy industry own use", neg_supply_in_fd = c("Exports",
  "International aviation bunkers", "International marine bunkers",
  "Losses", "Statistical differences", "Stock changes"),
  matname = "matname", U_excl_EIOU = "U_excl_EIOU",
  U_EIOU = "U_EIOU", V = "V", Y = "Y")

Arguments

.DF
a data frame with ledger_side, energy, flow_aggregation_point, and flow columns.

ledger_side
the name of the column in .DF that contains ledger side (a string). Default is "Ledger.side".

energy
the name of the column in .DF that contains energy and exergy values (a string). Default is "EX.ktoe".

flow_aggregation_point
the name of the column in .DF that contains flow aggregation point information. Default is "Flow.aggregation.point".

flow
the name of the column in .DF that contains flow information. Default is "Flow".

supply_side
the identifier for items on the supply side of the ledger (a string). Default is "Supply".

consumption_side
the identifier for items on the consumption side of the ledger (a string). Default is "Consumption".

eiou
the identifier for items that are energy industry own use. Default is ""Energy industry own use".

neg_supply_in_fd
identifiers for flow items that, when negative, are entries in the final demand (Y) matrix.

matname
the name of the output column containing the name of the matrix to which a row's value belongs (a string). Default is "matname".

U_excl_EIOU
the name for the use matrix that excludes energy industry own use (a string). Default is "U_excl_EIOU".

U_EIOU
the name for the energy industry own use matrix. Default is "U_EIOU".

V
the name for the make matrix (a string). Default is "V".

Y
the name for the final demand matrix (a string). Default is "Y".

Value

.DF with an added column, matname.

Details

In a reasonable workflow, this function would be followed by a call to add_row_col_meta and collapse_to_matrices.

Examples

add_matnames_iea(UKEnergy2000tidy)

#>     Country Year Ledger.side      Flow.aggregation.point Energy.type Last.stage
#> 1        GB 2000      Supply Total primary energy supply      E.ktoe      final
#> 2        GB 2000      Supply Total primary energy supply      E.ktoe      final
#> 3        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 4        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 5        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 6        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 7        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 8        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 9        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 10       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 11       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 12       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 13       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 14       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 15       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 16       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 17       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 18       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 19       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 20       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 21       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 22       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 23       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 24       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 25       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 26       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 27       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 28       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 29       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 30       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 31       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 32       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 33       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 34       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 35       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 36       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 37       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 38       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 39       GB 2000 Consumption                 Residential      E.ktoe      final
#> 40       GB 2000 Consumption                   Transport      E.ktoe      final
#> 41       GB 2000 Consumption                 Residential      E.ktoe      final
#> 42       GB 2000 Consumption                   Transport      E.ktoe      final
#> 43       GB 2000      Supply Total primary energy supply      E.ktoe     useful
#> 44       GB 2000      Supply Total primary energy supply      E.ktoe     useful
#> 45       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 46       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 47       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 48       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 49       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 50       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 51       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 52       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 53       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 54       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 55       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 56       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 57       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 58       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 59       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 60       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 61       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 62       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 63       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 64       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 65       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 66       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 67       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 68       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 69       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 70       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 71       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 72       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 73       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 74       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 75       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 76       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 77       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 78       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 79       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 80       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 81       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 82       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 83       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 84       GB 2000 Consumption                 Residential      E.ktoe     useful
#> 85       GB 2000 Consumption                   Transport      E.ktoe     useful
#> 86       GB 2000 Consumption                 Residential      E.ktoe     useful
#> 87       GB 2000 Consumption                   Transport      E.ktoe     useful
#> 88       GB 2000      Supply Total primary energy supply      E.ktoe   services
#> 89       GB 2000      Supply Total primary energy supply      E.ktoe   services
#> 90       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 91       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 92       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 93       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 94       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 95       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 96       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 97       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 98       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 99       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 100      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 101      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 102      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 103      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 104      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 105      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 106      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 107      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 108      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 109      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 110      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 111      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 112      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 113      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 114      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 115      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 116      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 117      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 118      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 119      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 120      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 121      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 122      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 123      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 124      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 125      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 126      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 127      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 128      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 129      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 130      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 131      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 132      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 133      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 134      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 135      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 136      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 137      GB 2000 Consumption                   Transport      E.ktoe   services
#> 138      GB 2000 Consumption                 Residential      E.ktoe   services
#> 139      GB 2000 Consumption                   Transport      E.ktoe   services
#> 140      GB 2000 Consumption                 Residential      E.ktoe   services
#> 141      GB 2000      Supply Total primary energy supply      X.ktoe   services
#> 142      GB 2000      Supply Total primary energy supply      X.ktoe   services
#> 143      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 144      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 145      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 146      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 147      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 148      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 149      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 150      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 151      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 152      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 153      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 154      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 155      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 156      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 157      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 158      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 159      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 160      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 161      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 162      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 163      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 164      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 165      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 166      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 167      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 168      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 169      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 170      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 171      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 172      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 173      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 174      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 175      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 176      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 177      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 178      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 179      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 180      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 181      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 182      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 183      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 184      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 185      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 186      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 187      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 188      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 189      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 190      GB 2000 Consumption                   Transport      X.ktoe   services
#> 191      GB 2000 Consumption                 Residential      X.ktoe   services
#> 192      GB 2000 Consumption                   Transport      X.ktoe   services
#> 193      GB 2000 Consumption                 Residential      X.ktoe   services
#>                  Flow                     Product       EX.ktoe            Unit
#> 1   Resources - Crude                       Crude  5.000000e+04            ktoe
#> 2      Resources - NG                          NG  4.300000e+04            ktoe
#> 3   Gas wells & proc.                  NG - Wells  4.300000e+04            ktoe
#> 4          Oil fields              Crude - Fields  5.000000e+04            ktoe
#> 5         Crude dist.               Crude - Dist.  4.750000e+04            ktoe
#> 6            NG dist.                  NG - Dist.  4.100000e+04            ktoe
#> 7      Oil refineries                      Diesel  2.050000e+04            ktoe
#> 8      Oil refineries                      Petrol  2.650000e+04            ktoe
#> 9        Power plants                       Elect  6.400000e+03            ktoe
#> 10        Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 11       Diesel dist.              Diesel - Dist.  1.550000e+04            ktoe
#> 12       Petrol dist.              Petrol - Dist.  2.650000e+04            ktoe
#> 13         Oil fields                       Crude -5.000000e+04            ktoe
#> 14        Crude dist.              Crude - Fields -4.750000e+04            ktoe
#> 15     Oil refineries               Crude - Dist. -4.700000e+04            ktoe
#> 16  Gas wells & proc.                          NG -4.300000e+04            ktoe
#> 17           NG dist.                  NG - Wells -4.100000e+04            ktoe
#> 18       Power plants                  NG - Dist. -1.600000e+04            ktoe
#> 19       Diesel dist.                      Diesel -1.550000e+04            ktoe
#> 20        Elect. grid                       Elect -6.400000e+03            ktoe
#> 21       Petrol dist.                      Petrol -2.650000e+04            ktoe
#> 22  Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 23         Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 24        Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 25           NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 26     Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 27       Power plants                Elect - Grid -1.000000e+02            ktoe
#> 28     Oil refineries                      Diesel -5.000000e+03            ktoe
#> 29        Crude dist.               Crude - Dist. -5.000000e+02            ktoe
#> 30  Gas wells & proc.              Diesel - Dist. -5.000000e+01            ktoe
#> 31         Oil fields              Diesel - Dist. -5.000000e+01            ktoe
#> 32        Crude dist.              Diesel - Dist. -2.500000e+01            ktoe
#> 33           NG dist.              Diesel - Dist. -2.500000e+01            ktoe
#> 34       Petrol dist.              Diesel - Dist. -2.500000e+02            ktoe
#> 35  Gas wells & proc.                  NG - Wells -2.000000e+03            ktoe
#> 36         Oil fields              Crude - Fields -2.500000e+03            ktoe
#> 37       Diesel dist.              Diesel - Dist. -3.500000e+02            ktoe
#> 38       Petrol dist.              Petrol - Dist. -5.000000e+02            ktoe
#> 39        Residential                  NG - Dist.  2.500000e+04            ktoe
#> 40          Transport              Diesel - Dist.  1.475000e+04            ktoe
#> 41        Residential                Elect - Grid  6.000000e+03            ktoe
#> 42          Transport              Petrol - Dist.  2.600000e+04            ktoe
#> 43  Resources - Crude                       Crude  5.000000e+04            ktoe
#> 44     Resources - NG                          NG  4.300000e+04            ktoe
#> 45  Gas wells & proc.                  NG - Wells  4.100000e+04            ktoe
#> 46         Oil fields              Crude - Fields  4.750000e+04            ktoe
#> 47        Crude dist.               Crude - Dist.  4.700000e+04            ktoe
#> 48           NG dist.                  NG - Dist.  4.100000e+04            ktoe
#> 49     Oil refineries                      Diesel  1.550000e+04            ktoe
#> 50     Oil refineries                      Petrol  2.650000e+04            ktoe
#> 51       Power plants                       Elect  6.400000e+03            ktoe
#> 52        Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 53       Diesel dist.              Diesel - Dist.  1.515000e+04            ktoe
#> 54       Petrol dist.              Petrol - Dist.  2.600000e+04            ktoe
#> 55        Car engines            MD - Car engines  3.000400e+03            ktoe
#> 56           Furnaces                         LTH  2.000000e+04            ktoe
#> 57     Light fixtures                       Light  1.200000e+03            ktoe
#> 58      Truck engines          MD - Truck engines  1.799980e+03            ktoe
#> 59         Oil fields                       Crude -5.000000e+04            ktoe
#> 60        Crude dist.              Crude - Fields -4.750000e+04            ktoe
#> 61     Oil refineries               Crude - Dist. -4.700000e+04            ktoe
#> 62  Gas wells & proc.                          NG -4.300000e+04            ktoe
#> 63           NG dist.                  NG - Wells -4.100000e+04            ktoe
#> 64       Power plants                  NG - Dist. -1.600000e+04            ktoe
#> 65           Furnaces                  NG - Dist. -2.500000e+04            ktoe
#> 66       Diesel dist.                      Diesel -1.550000e+04            ktoe
#> 67  Gas wells & proc.              Diesel - Dist. -5.000000e+01            ktoe
#> 68         Oil fields              Diesel - Dist. -5.000000e+01            ktoe
#> 69      Truck engines              Diesel - Dist. -1.505000e+04            ktoe
#> 70        Elect. grid                       Elect -6.400000e+03            ktoe
#> 71  Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 72         Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 73        Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 74           NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 75     Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 76       Power plants                Elect - Grid -1.000000e+02            ktoe
#> 77     Light fixtures                Elect - Grid -6.000000e+03            ktoe
#> 78       Petrol dist.                      Petrol -2.650000e+04            ktoe
#> 79        Car engines              Petrol - Dist. -2.600000e+04            ktoe
#> 80        Crude dist.          MD - Truck engines -2.000000e+01            ktoe
#> 81           NG dist.          MD - Truck engines -2.000000e+01            ktoe
#> 82       Diesel dist.          MD - Truck engines -1.799980e+01            ktoe
#> 83       Petrol dist.          MD - Truck engines -2.699970e+01            ktoe
#> 84        Residential                       Light  1.200000e+03            ktoe
#> 85          Transport                         LTH  2.000000e+04            ktoe
#> 86        Residential            MD - Car engines  3.000400e+03            ktoe
#> 87          Transport          MD - Truck engines  1.714980e+03            ktoe
#> 88  Resources - Crude                       Crude  5.000000e+04            ktoe
#> 89     Resources - NG                          NG  4.300000e+04            ktoe
#> 90  Gas wells & proc.                  NG - Wells  4.100000e+04            ktoe
#> 91         Oil fields              Crude - Fields  4.750000e+04            ktoe
#> 92        Crude dist.               Crude - Dist.  4.700000e+04            ktoe
#> 93           NG dist.                  NG - Dist.  4.100000e+04            ktoe
#> 94     Oil refineries                      Diesel  1.550000e+04            ktoe
#> 95     Oil refineries                      Petrol  2.650000e+04            ktoe
#> 96       Power plants                       Elect  6.400000e+03            ktoe
#> 97        Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 98       Diesel dist.              Diesel - Dist.  1.515000e+04            ktoe
#> 99       Petrol dist.              Petrol - Dist.  2.600000e+04            ktoe
#> 100       Car engines            MD - Car engines  3.000400e+03            ktoe
#> 101          Furnaces                         LTH  2.000000e+04            ktoe
#> 102    Light fixtures                       Light  1.200000e+03            ktoe
#> 103     Truck engines          MD - Truck engines  1.799980e+03            ktoe
#> 104              Cars Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 105             Homes        Space heating [m3-K]  7.500000e+10            m3-K
#> 106             Rooms Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 107            Trucks     Freight [tonne-km/year]  1.500000e+11     tonne-km/yr
#> 108        Oil fields                       Crude -5.000000e+04            ktoe
#> 109       Crude dist.              Crude - Fields -4.750000e+04            ktoe
#> 110    Oil refineries               Crude - Dist. -4.700000e+04            ktoe
#> 111 Gas wells & proc.                          NG -4.300000e+04            ktoe
#> 112          NG dist.                  NG - Wells -4.100000e+04            ktoe
#> 113      Power plants                  NG - Dist. -1.600000e+04            ktoe
#> 114          Furnaces                  NG - Dist. -2.500000e+04            ktoe
#> 115      Diesel dist.                      Diesel -1.550000e+04            ktoe
#> 116 Gas wells & proc.              Diesel - Dist. -5.000000e+01            ktoe
#> 117        Oil fields              Diesel - Dist. -5.000000e+01            ktoe
#> 118     Truck engines              Diesel - Dist. -1.505000e+04            ktoe
#> 119       Elect. grid                       Elect -6.400000e+03            ktoe
#> 120 Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 121        Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 122       Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 123          NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 124    Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 125      Power plants                Elect - Grid -1.000000e+02            ktoe
#> 126    Light fixtures                Elect - Grid -6.000000e+03            ktoe
#> 127      Petrol dist.                      Petrol -2.650000e+04            ktoe
#> 128       Car engines              Petrol - Dist. -2.600000e+04            ktoe
#> 129             Rooms                       Light -1.200000e+03            ktoe
#> 130             Homes                         LTH -2.000000e+04            ktoe
#> 131              Cars            MD - Car engines -3.000400e+03            ktoe
#> 132            Trucks          MD - Truck engines -1.799980e+03            ktoe
#> 133       Crude dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 134          NG dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 135      Diesel dist.     Freight [tonne-km/year] -1.500000e+09     tonne-km/yr
#> 136      Petrol dist.     Freight [tonne-km/year] -2.250000e+09     tonne-km/yr
#> 137         Transport     Freight [tonne-km/year]  1.429166e+11     tonne-km/yr
#> 138       Residential Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 139         Transport Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 140       Residential        Space heating [m3-K]  7.500000e+10            m3-K
#> 141 Resources - Crude                       Crude  5.350000e+04            ktoe
#> 142    Resources - NG                          NG  4.472000e+04            ktoe
#> 143 Gas wells & proc.                  NG - Wells  4.264000e+04            ktoe
#> 144        Oil fields              Crude - Fields  5.082500e+04            ktoe
#> 145       Crude dist.               Crude - Dist.  5.029000e+04            ktoe
#> 146          NG dist.                  NG - Dist.  4.264000e+04            ktoe
#> 147    Oil refineries                      Diesel  1.658500e+04            ktoe
#> 148    Oil refineries                      Petrol  2.835500e+04            ktoe
#> 149      Power plants                       Elect  6.400000e+03            ktoe
#> 150       Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 151      Diesel dist.              Diesel - Dist.  1.621050e+04            ktoe
#> 152      Petrol dist.              Petrol - Dist.  2.782000e+04            ktoe
#> 153       Car engines            MD - Car engines  3.000400e+03            ktoe
#> 154          Furnaces                         LTH  3.354016e+02            ktoe
#> 155    Light fixtures                       Light  9.600000e+01            ktoe
#> 156     Truck engines          MD - Truck engines  1.799980e+03            ktoe
#> 157              Cars Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 158             Homes        Space heating [m3-K]  7.500000e+10            m3-K
#> 159             Rooms Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 160            Trucks     Freight [tonne-km/year]  1.500000e+11     tonne-km/yr
#> 161        Oil fields                       Crude -5.350000e+04            ktoe
#> 162       Crude dist.              Crude - Fields -5.082500e+04            ktoe
#> 163    Oil refineries               Crude - Dist. -5.029000e+04            ktoe
#> 164 Gas wells & proc.                          NG -4.472000e+04            ktoe
#> 165          NG dist.                  NG - Wells -4.264000e+04            ktoe
#> 166      Power plants                  NG - Dist. -1.664000e+04            ktoe
#> 167          Furnaces                  NG - Dist. -2.600000e+04            ktoe
#> 168      Diesel dist.                      Diesel -1.658500e+04            ktoe
#> 169 Gas wells & proc.              Diesel - Dist. -5.350000e+01            ktoe
#> 170        Oil fields              Diesel - Dist. -5.350000e+01            ktoe
#> 171     Truck engines              Diesel - Dist. -1.610350e+04            ktoe
#> 172       Elect. grid                       Elect -6.400000e+03            ktoe
#> 173 Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 174        Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 175       Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 176          NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 177    Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 178      Power plants                Elect - Grid -1.000000e+02            ktoe
#> 179    Light fixtures                Elect - Grid -6.000000e+03            ktoe
#> 180      Petrol dist.                      Petrol -2.835500e+04            ktoe
#> 181       Car engines              Petrol - Dist. -2.782000e+04            ktoe
#> 182             Rooms                       Light -9.600000e+01            ktoe
#> 183             Homes                         LTH -3.354016e+02            ktoe
#> 184              Cars            MD - Car engines -3.000400e+03            ktoe
#> 185            Trucks          MD - Truck engines -1.799980e+03            ktoe
#> 186       Crude dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 187          NG dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 188      Diesel dist.     Freight [tonne-km/year] -1.500000e+09     tonne-km/yr
#> 189      Petrol dist.     Freight [tonne-km/year] -2.250000e+09     tonne-km/yr
#> 190         Transport     Freight [tonne-km/year]  1.429166e+11     tonne-km/yr
#> 191       Residential Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 192         Transport Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 193       Residential        Space heating [m3-K]  7.500000e+10            m3-K
#>         matname
#> 1             V
#> 2             V
#> 3             V
#> 4             V
#> 5             V
#> 6             V
#> 7             V
#> 8             V
#> 9             V
#> 10            V
#> 11            V
#> 12            V
#> 13  U_excl_EIOU
#> 14  U_excl_EIOU
#> 15  U_excl_EIOU
#> 16  U_excl_EIOU
#> 17  U_excl_EIOU
#> 18  U_excl_EIOU
#> 19  U_excl_EIOU
#> 20  U_excl_EIOU
#> 21  U_excl_EIOU
#> 22       U_EIOU
#> 23       U_EIOU
#> 24       U_EIOU
#> 25       U_EIOU
#> 26       U_EIOU
#> 27       U_EIOU
#> 28       U_EIOU
#> 29       U_EIOU
#> 30       U_EIOU
#> 31       U_EIOU
#> 32       U_EIOU
#> 33       U_EIOU
#> 34       U_EIOU
#> 35       U_EIOU
#> 36       U_EIOU
#> 37       U_EIOU
#> 38       U_EIOU
#> 39            Y
#> 40            Y
#> 41            Y
#> 42            Y
#> 43            V
#> 44            V
#> 45            V
#> 46            V
#> 47            V
#> 48            V
#> 49            V
#> 50            V
#> 51            V
#> 52            V
#> 53            V
#> 54            V
#> 55            V
#> 56            V
#> 57            V
#> 58            V
#> 59  U_excl_EIOU
#> 60  U_excl_EIOU
#> 61  U_excl_EIOU
#> 62  U_excl_EIOU
#> 63  U_excl_EIOU
#> 64  U_excl_EIOU
#> 65  U_excl_EIOU
#> 66  U_excl_EIOU
#> 67       U_EIOU
#> 68       U_EIOU
#> 69  U_excl_EIOU
#> 70  U_excl_EIOU
#> 71       U_EIOU
#> 72       U_EIOU
#> 73       U_EIOU
#> 74       U_EIOU
#> 75       U_EIOU
#> 76       U_EIOU
#> 77  U_excl_EIOU
#> 78  U_excl_EIOU
#> 79  U_excl_EIOU
#> 80       U_EIOU
#> 81       U_EIOU
#> 82       U_EIOU
#> 83       U_EIOU
#> 84            Y
#> 85            Y
#> 86            Y
#> 87            Y
#> 88            V
#> 89            V
#> 90            V
#> 91            V
#> 92            V
#> 93            V
#> 94            V
#> 95            V
#> 96            V
#> 97            V
#> 98            V
#> 99            V
#> 100           V
#> 101           V
#> 102           V
#> 103           V
#> 104           V
#> 105           V
#> 106           V
#> 107           V
#> 108 U_excl_EIOU
#> 109 U_excl_EIOU
#> 110 U_excl_EIOU
#> 111 U_excl_EIOU
#> 112 U_excl_EIOU
#> 113 U_excl_EIOU
#> 114 U_excl_EIOU
#> 115 U_excl_EIOU
#> 116      U_EIOU
#> 117      U_EIOU
#> 118 U_excl_EIOU
#> 119 U_excl_EIOU
#> 120      U_EIOU
#> 121      U_EIOU
#> 122      U_EIOU
#> 123      U_EIOU
#> 124      U_EIOU
#> 125      U_EIOU
#> 126 U_excl_EIOU
#> 127 U_excl_EIOU
#> 128 U_excl_EIOU
#> 129 U_excl_EIOU
#> 130 U_excl_EIOU
#> 131 U_excl_EIOU
#> 132 U_excl_EIOU
#> 133      U_EIOU
#> 134      U_EIOU
#> 135      U_EIOU
#> 136      U_EIOU
#> 137           Y
#> 138           Y
#> 139           Y
#> 140           Y
#> 141           V
#> 142           V
#> 143           V
#> 144           V
#> 145           V
#> 146           V
#> 147           V
#> 148           V
#> 149           V
#> 150           V
#> 151           V
#> 152           V
#> 153           V
#> 154           V
#> 155           V
#> 156           V
#> 157           V
#> 158           V
#> 159           V
#> 160           V
#> 161 U_excl_EIOU
#> 162 U_excl_EIOU
#> 163 U_excl_EIOU
#> 164 U_excl_EIOU
#> 165 U_excl_EIOU
#> 166 U_excl_EIOU
#> 167 U_excl_EIOU
#> 168 U_excl_EIOU
#> 169      U_EIOU
#> 170      U_EIOU
#> 171 U_excl_EIOU
#> 172 U_excl_EIOU
#> 173      U_EIOU
#> 174      U_EIOU
#> 175      U_EIOU
#> 176      U_EIOU
#> 177      U_EIOU
#> 178      U_EIOU
#> 179 U_excl_EIOU
#> 180 U_excl_EIOU
#> 181 U_excl_EIOU
#> 182 U_excl_EIOU
#> 183 U_excl_EIOU
#> 184 U_excl_EIOU
#> 185 U_excl_EIOU
#> 186      U_EIOU
#> 187      U_EIOU
#> 188      U_EIOU
#> 189      U_EIOU
#> 190           Y
#> 191           Y
#> 192           Y
#> 193           Y