Add row, column, row type, and column type metadata

Add row, column, row type, and column type metadata

add_row_col_meta(.DF, matname = "matname", product = "Product",
  flow = "Flow", U = "U", U_EIOU = "U_EIOU", R = "R", V = "V",
  Y = "Y", industry_type = "Industry", product_type = "Product",
  sector_type = "Industry", resource_type = "Industry",
  rowname = "rowname", colname = "colname", rowtype = "rowtype",
  coltype = "coltype")

Arguments

.DF
a data frame containing matname_colname.

matname
the name of the column in .DF that contains names of matrices (a string). Default is "matname".

product
the name of the column in .DF where Product names is found (a string). Default is "Product".

flow
the name of the column in .DF where Flow names is found (a string). Default is "Flow".

U
the name for use matrices (a string). Default is "U".

U_EIOU
the name for energy industry own use matrices (a string). Default is "U_EIOU".

R
the name for resource matrices (a string). Default is "R".

V
the name for make matrices (a string). Default is "V".

Y
the name for final demand matrices (a string). Default is "Y".

industry_type
the name that identifies production industries and and transformation processes (a string). Default is "Industry".

product_type
the name that identifies energy carriers (a string). Default is "Product".

sector_type
the name that identifies final demand sectors (a string). Default is "Industry".

resource_type
the name that identifies resource sectors (a string). Default is "Industry".

rowname
the name of the output column that contains row names for matrices (a string). Default is "rowname".

colname
the name of the output column that contains column names for matrices (a string). Default is "colname".

rowtype
the name of the output column that contains row types for matrices (a string). Default is "rowtype".

coltype
the name of the output column that contains column types for matrices (a string). Default is "coltype".

Value

.DF with additional columns named rowname, colname, rowtype, and coltype.

Examples

UKEnergy2000tidy %>%
  add_matnames_iea() %>%
  add_row_col_meta()

#>     Country Year Ledger.side      Flow.aggregation.point Energy.type Last.stage
#> 1        GB 2000      Supply Total primary energy supply      E.ktoe      final
#> 2        GB 2000      Supply Total primary energy supply      E.ktoe      final
#> 3        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 4        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 5        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 6        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 7        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 8        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 9        GB 2000      Supply    Transformation processes      E.ktoe      final
#> 10       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 11       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 12       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 13       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 14       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 15       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 16       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 17       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 18       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 19       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 20       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 21       GB 2000      Supply    Transformation processes      E.ktoe      final
#> 22       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 23       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 24       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 25       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 26       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 27       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 28       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 29       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 30       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 31       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 32       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 33       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 34       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 35       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 36       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 37       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 38       GB 2000      Supply     Energy industry own use      E.ktoe      final
#> 39       GB 2000 Consumption                 Residential      E.ktoe      final
#> 40       GB 2000 Consumption                   Transport      E.ktoe      final
#> 41       GB 2000 Consumption                 Residential      E.ktoe      final
#> 42       GB 2000 Consumption                   Transport      E.ktoe      final
#> 43       GB 2000      Supply Total primary energy supply      E.ktoe     useful
#> 44       GB 2000      Supply Total primary energy supply      E.ktoe     useful
#> 45       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 46       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 47       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 48       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 49       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 50       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 51       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 52       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 53       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 54       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 55       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 56       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 57       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 58       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 59       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 60       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 61       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 62       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 63       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 64       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 65       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 66       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 67       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 68       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 69       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 70       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 71       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 72       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 73       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 74       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 75       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 76       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 77       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 78       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 79       GB 2000      Supply    Transformation processes      E.ktoe     useful
#> 80       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 81       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 82       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 83       GB 2000      Supply     Energy industry own use      E.ktoe     useful
#> 84       GB 2000 Consumption                 Residential      E.ktoe     useful
#> 85       GB 2000 Consumption                   Transport      E.ktoe     useful
#> 86       GB 2000 Consumption                 Residential      E.ktoe     useful
#> 87       GB 2000 Consumption                   Transport      E.ktoe     useful
#> 88       GB 2000      Supply Total primary energy supply      E.ktoe   services
#> 89       GB 2000      Supply Total primary energy supply      E.ktoe   services
#> 90       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 91       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 92       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 93       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 94       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 95       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 96       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 97       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 98       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 99       GB 2000      Supply    Transformation processes      E.ktoe   services
#> 100      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 101      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 102      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 103      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 104      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 105      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 106      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 107      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 108      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 109      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 110      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 111      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 112      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 113      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 114      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 115      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 116      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 117      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 118      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 119      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 120      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 121      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 122      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 123      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 124      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 125      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 126      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 127      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 128      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 129      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 130      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 131      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 132      GB 2000      Supply    Transformation processes      E.ktoe   services
#> 133      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 134      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 135      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 136      GB 2000      Supply     Energy industry own use      E.ktoe   services
#> 137      GB 2000 Consumption                   Transport      E.ktoe   services
#> 138      GB 2000 Consumption                 Residential      E.ktoe   services
#> 139      GB 2000 Consumption                   Transport      E.ktoe   services
#> 140      GB 2000 Consumption                 Residential      E.ktoe   services
#> 141      GB 2000      Supply Total primary energy supply      X.ktoe   services
#> 142      GB 2000      Supply Total primary energy supply      X.ktoe   services
#> 143      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 144      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 145      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 146      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 147      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 148      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 149      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 150      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 151      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 152      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 153      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 154      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 155      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 156      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 157      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 158      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 159      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 160      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 161      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 162      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 163      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 164      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 165      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 166      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 167      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 168      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 169      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 170      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 171      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 172      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 173      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 174      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 175      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 176      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 177      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 178      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 179      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 180      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 181      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 182      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 183      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 184      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 185      GB 2000      Supply    Transformation processes      X.ktoe   services
#> 186      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 187      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 188      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 189      GB 2000      Supply     Energy industry own use      X.ktoe   services
#> 190      GB 2000 Consumption                   Transport      X.ktoe   services
#> 191      GB 2000 Consumption                 Residential      X.ktoe   services
#> 192      GB 2000 Consumption                   Transport      X.ktoe   services
#> 193      GB 2000 Consumption                 Residential      X.ktoe   services
#>                  Flow                     Product       EX.ktoe            Unit
#> 1   Resources - Crude                       Crude  5.000000e+04            ktoe
#> 2      Resources - NG                          NG  4.300000e+04            ktoe
#> 3   Gas wells & proc.                  NG - Wells  4.300000e+04            ktoe
#> 4          Oil fields              Crude - Fields  5.000000e+04            ktoe
#> 5         Crude dist.               Crude - Dist.  4.750000e+04            ktoe
#> 6            NG dist.                  NG - Dist.  4.100000e+04            ktoe
#> 7      Oil refineries                      Diesel  2.050000e+04            ktoe
#> 8      Oil refineries                      Petrol  2.650000e+04            ktoe
#> 9        Power plants                       Elect  6.400000e+03            ktoe
#> 10        Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 11       Diesel dist.              Diesel - Dist.  1.550000e+04            ktoe
#> 12       Petrol dist.              Petrol - Dist.  2.650000e+04            ktoe
#> 13         Oil fields                       Crude -5.000000e+04            ktoe
#> 14        Crude dist.              Crude - Fields -4.750000e+04            ktoe
#> 15     Oil refineries               Crude - Dist. -4.700000e+04            ktoe
#> 16  Gas wells & proc.                          NG -4.300000e+04            ktoe
#> 17           NG dist.                  NG - Wells -4.100000e+04            ktoe
#> 18       Power plants                  NG - Dist. -1.600000e+04            ktoe
#> 19       Diesel dist.                      Diesel -1.550000e+04            ktoe
#> 20        Elect. grid                       Elect -6.400000e+03            ktoe
#> 21       Petrol dist.                      Petrol -2.650000e+04            ktoe
#> 22  Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 23         Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 24        Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 25           NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 26     Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 27       Power plants                Elect - Grid -1.000000e+02            ktoe
#> 28     Oil refineries                      Diesel -5.000000e+03            ktoe
#> 29        Crude dist.               Crude - Dist. -5.000000e+02            ktoe
#> 30  Gas wells & proc.              Diesel - Dist. -5.000000e+01            ktoe
#> 31         Oil fields              Diesel - Dist. -5.000000e+01            ktoe
#> 32        Crude dist.              Diesel - Dist. -2.500000e+01            ktoe
#> 33           NG dist.              Diesel - Dist. -2.500000e+01            ktoe
#> 34       Petrol dist.              Diesel - Dist. -2.500000e+02            ktoe
#> 35  Gas wells & proc.                  NG - Wells -2.000000e+03            ktoe
#> 36         Oil fields              Crude - Fields -2.500000e+03            ktoe
#> 37       Diesel dist.              Diesel - Dist. -3.500000e+02            ktoe
#> 38       Petrol dist.              Petrol - Dist. -5.000000e+02            ktoe
#> 39        Residential                  NG - Dist.  2.500000e+04            ktoe
#> 40          Transport              Diesel - Dist.  1.475000e+04            ktoe
#> 41        Residential                Elect - Grid  6.000000e+03            ktoe
#> 42          Transport              Petrol - Dist.  2.600000e+04            ktoe
#> 43  Resources - Crude                       Crude  5.000000e+04            ktoe
#> 44     Resources - NG                          NG  4.300000e+04            ktoe
#> 45  Gas wells & proc.                  NG - Wells  4.100000e+04            ktoe
#> 46         Oil fields              Crude - Fields  4.750000e+04            ktoe
#> 47        Crude dist.               Crude - Dist.  4.700000e+04            ktoe
#> 48           NG dist.                  NG - Dist.  4.100000e+04            ktoe
#> 49     Oil refineries                      Diesel  1.550000e+04            ktoe
#> 50     Oil refineries                      Petrol  2.650000e+04            ktoe
#> 51       Power plants                       Elect  6.400000e+03            ktoe
#> 52        Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 53       Diesel dist.              Diesel - Dist.  1.515000e+04            ktoe
#> 54       Petrol dist.              Petrol - Dist.  2.600000e+04            ktoe
#> 55        Car engines            MD - Car engines  3.000400e+03            ktoe
#> 56           Furnaces                         LTH  2.000000e+04            ktoe
#> 57     Light fixtures                       Light  1.200000e+03            ktoe
#> 58      Truck engines          MD - Truck engines  1.799980e+03            ktoe
#> 59         Oil fields                       Crude -5.000000e+04            ktoe
#> 60        Crude dist.              Crude - Fields -4.750000e+04            ktoe
#> 61     Oil refineries               Crude - Dist. -4.700000e+04            ktoe
#> 62  Gas wells & proc.                          NG -4.300000e+04            ktoe
#> 63           NG dist.                  NG - Wells -4.100000e+04            ktoe
#> 64       Power plants                  NG - Dist. -1.600000e+04            ktoe
#> 65           Furnaces                  NG - Dist. -2.500000e+04            ktoe
#> 66       Diesel dist.                      Diesel -1.550000e+04            ktoe
#> 67  Gas wells & proc.              Diesel - Dist. -5.000000e+01            ktoe
#> 68         Oil fields              Diesel - Dist. -5.000000e+01            ktoe
#> 69      Truck engines              Diesel - Dist. -1.505000e+04            ktoe
#> 70        Elect. grid                       Elect -6.400000e+03            ktoe
#> 71  Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 72         Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 73        Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 74           NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 75     Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 76       Power plants                Elect - Grid -1.000000e+02            ktoe
#> 77     Light fixtures                Elect - Grid -6.000000e+03            ktoe
#> 78       Petrol dist.                      Petrol -2.650000e+04            ktoe
#> 79        Car engines              Petrol - Dist. -2.600000e+04            ktoe
#> 80        Crude dist.          MD - Truck engines -2.000000e+01            ktoe
#> 81           NG dist.          MD - Truck engines -2.000000e+01            ktoe
#> 82       Diesel dist.          MD - Truck engines -1.799980e+01            ktoe
#> 83       Petrol dist.          MD - Truck engines -2.699970e+01            ktoe
#> 84        Residential                       Light  1.200000e+03            ktoe
#> 85          Transport                         LTH  2.000000e+04            ktoe
#> 86        Residential            MD - Car engines  3.000400e+03            ktoe
#> 87          Transport          MD - Truck engines  1.714980e+03            ktoe
#> 88  Resources - Crude                       Crude  5.000000e+04            ktoe
#> 89     Resources - NG                          NG  4.300000e+04            ktoe
#> 90  Gas wells & proc.                  NG - Wells  4.100000e+04            ktoe
#> 91         Oil fields              Crude - Fields  4.750000e+04            ktoe
#> 92        Crude dist.               Crude - Dist.  4.700000e+04            ktoe
#> 93           NG dist.                  NG - Dist.  4.100000e+04            ktoe
#> 94     Oil refineries                      Diesel  1.550000e+04            ktoe
#> 95     Oil refineries                      Petrol  2.650000e+04            ktoe
#> 96       Power plants                       Elect  6.400000e+03            ktoe
#> 97        Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 98       Diesel dist.              Diesel - Dist.  1.515000e+04            ktoe
#> 99       Petrol dist.              Petrol - Dist.  2.600000e+04            ktoe
#> 100       Car engines            MD - Car engines  3.000400e+03            ktoe
#> 101          Furnaces                         LTH  2.000000e+04            ktoe
#> 102    Light fixtures                       Light  1.200000e+03            ktoe
#> 103     Truck engines          MD - Truck engines  1.799980e+03            ktoe
#> 104              Cars Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 105             Homes        Space heating [m3-K]  7.500000e+10            m3-K
#> 106             Rooms Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 107            Trucks     Freight [tonne-km/year]  1.500000e+11     tonne-km/yr
#> 108        Oil fields                       Crude -5.000000e+04            ktoe
#> 109       Crude dist.              Crude - Fields -4.750000e+04            ktoe
#> 110    Oil refineries               Crude - Dist. -4.700000e+04            ktoe
#> 111 Gas wells & proc.                          NG -4.300000e+04            ktoe
#> 112          NG dist.                  NG - Wells -4.100000e+04            ktoe
#> 113      Power plants                  NG - Dist. -1.600000e+04            ktoe
#> 114          Furnaces                  NG - Dist. -2.500000e+04            ktoe
#> 115      Diesel dist.                      Diesel -1.550000e+04            ktoe
#> 116 Gas wells & proc.              Diesel - Dist. -5.000000e+01            ktoe
#> 117        Oil fields              Diesel - Dist. -5.000000e+01            ktoe
#> 118     Truck engines              Diesel - Dist. -1.505000e+04            ktoe
#> 119       Elect. grid                       Elect -6.400000e+03            ktoe
#> 120 Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 121        Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 122       Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 123          NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 124    Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 125      Power plants                Elect - Grid -1.000000e+02            ktoe
#> 126    Light fixtures                Elect - Grid -6.000000e+03            ktoe
#> 127      Petrol dist.                      Petrol -2.650000e+04            ktoe
#> 128       Car engines              Petrol - Dist. -2.600000e+04            ktoe
#> 129             Rooms                       Light -1.200000e+03            ktoe
#> 130             Homes                         LTH -2.000000e+04            ktoe
#> 131              Cars            MD - Car engines -3.000400e+03            ktoe
#> 132            Trucks          MD - Truck engines -1.799980e+03            ktoe
#> 133       Crude dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 134          NG dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 135      Diesel dist.     Freight [tonne-km/year] -1.500000e+09     tonne-km/yr
#> 136      Petrol dist.     Freight [tonne-km/year] -2.250000e+09     tonne-km/yr
#> 137         Transport     Freight [tonne-km/year]  1.429166e+11     tonne-km/yr
#> 138       Residential Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 139         Transport Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 140       Residential        Space heating [m3-K]  7.500000e+10            m3-K
#> 141 Resources - Crude                       Crude  5.350000e+04            ktoe
#> 142    Resources - NG                          NG  4.472000e+04            ktoe
#> 143 Gas wells & proc.                  NG - Wells  4.264000e+04            ktoe
#> 144        Oil fields              Crude - Fields  5.082500e+04            ktoe
#> 145       Crude dist.               Crude - Dist.  5.029000e+04            ktoe
#> 146          NG dist.                  NG - Dist.  4.264000e+04            ktoe
#> 147    Oil refineries                      Diesel  1.658500e+04            ktoe
#> 148    Oil refineries                      Petrol  2.835500e+04            ktoe
#> 149      Power plants                       Elect  6.400000e+03            ktoe
#> 150       Elect. grid                Elect - Grid  6.275000e+03            ktoe
#> 151      Diesel dist.              Diesel - Dist.  1.621050e+04            ktoe
#> 152      Petrol dist.              Petrol - Dist.  2.782000e+04            ktoe
#> 153       Car engines            MD - Car engines  3.000400e+03            ktoe
#> 154          Furnaces                         LTH  3.354016e+02            ktoe
#> 155    Light fixtures                       Light  9.600000e+01            ktoe
#> 156     Truck engines          MD - Truck engines  1.799980e+03            ktoe
#> 157              Cars Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 158             Homes        Space heating [m3-K]  7.500000e+10            m3-K
#> 159             Rooms Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 160            Trucks     Freight [tonne-km/year]  1.500000e+11     tonne-km/yr
#> 161        Oil fields                       Crude -5.350000e+04            ktoe
#> 162       Crude dist.              Crude - Fields -5.082500e+04            ktoe
#> 163    Oil refineries               Crude - Dist. -5.029000e+04            ktoe
#> 164 Gas wells & proc.                          NG -4.472000e+04            ktoe
#> 165          NG dist.                  NG - Wells -4.264000e+04            ktoe
#> 166      Power plants                  NG - Dist. -1.664000e+04            ktoe
#> 167          Furnaces                  NG - Dist. -2.600000e+04            ktoe
#> 168      Diesel dist.                      Diesel -1.658500e+04            ktoe
#> 169 Gas wells & proc.              Diesel - Dist. -5.350000e+01            ktoe
#> 170        Oil fields              Diesel - Dist. -5.350000e+01            ktoe
#> 171     Truck engines              Diesel - Dist. -1.610350e+04            ktoe
#> 172       Elect. grid                       Elect -6.400000e+03            ktoe
#> 173 Gas wells & proc.                Elect - Grid -2.500000e+01            ktoe
#> 174        Oil fields                Elect - Grid -2.500000e+01            ktoe
#> 175       Crude dist.                Elect - Grid -2.500000e+01            ktoe
#> 176          NG dist.                Elect - Grid -2.500000e+01            ktoe
#> 177    Oil refineries                Elect - Grid -7.500000e+01            ktoe
#> 178      Power plants                Elect - Grid -1.000000e+02            ktoe
#> 179    Light fixtures                Elect - Grid -6.000000e+03            ktoe
#> 180      Petrol dist.                      Petrol -2.835500e+04            ktoe
#> 181       Car engines              Petrol - Dist. -2.782000e+04            ktoe
#> 182             Rooms                       Light -9.600000e+01            ktoe
#> 183             Homes                         LTH -3.354016e+02            ktoe
#> 184              Cars            MD - Car engines -3.000400e+03            ktoe
#> 185            Trucks          MD - Truck engines -1.799980e+03            ktoe
#> 186       Crude dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 187          NG dist.     Freight [tonne-km/year] -1.666685e+09     tonne-km/yr
#> 188      Diesel dist.     Freight [tonne-km/year] -1.500000e+09     tonne-km/yr
#> 189      Petrol dist.     Freight [tonne-km/year] -2.250000e+09     tonne-km/yr
#> 190         Transport     Freight [tonne-km/year]  1.429166e+11     tonne-km/yr
#> 191       Residential Illumination [lumen-hrs/yr]  5.000000e+14    lumen-hrs/yr
#> 192         Transport Passenger [passenger-km/yr]  5.000000e+11 passenger-km/yr
#> 193       Residential        Space heating [m3-K]  7.500000e+10            m3-K
#>         matname                     rowname                     colname
#> 1             V           Resources - Crude                       Crude
#> 2             V              Resources - NG                          NG
#> 3             V           Gas wells & proc.                  NG - Wells
#> 4             V                  Oil fields              Crude - Fields
#> 5             V                 Crude dist.               Crude - Dist.
#> 6             V                    NG dist.                  NG - Dist.
#> 7             V              Oil refineries                      Diesel
#> 8             V              Oil refineries                      Petrol
#> 9             V                Power plants                       Elect
#> 10            V                 Elect. grid                Elect - Grid
#> 11            V                Diesel dist.              Diesel - Dist.
#> 12            V                Petrol dist.              Petrol - Dist.
#> 13  U_excl_EIOU                       Crude                  Oil fields
#> 14  U_excl_EIOU              Crude - Fields                 Crude dist.
#> 15  U_excl_EIOU               Crude - Dist.              Oil refineries
#> 16  U_excl_EIOU                          NG           Gas wells & proc.
#> 17  U_excl_EIOU                  NG - Wells                    NG dist.
#> 18  U_excl_EIOU                  NG - Dist.                Power plants
#> 19  U_excl_EIOU                      Diesel                Diesel dist.
#> 20  U_excl_EIOU                       Elect                 Elect. grid
#> 21  U_excl_EIOU                      Petrol                Petrol dist.
#> 22       U_EIOU                Elect - Grid           Gas wells & proc.
#> 23       U_EIOU                Elect - Grid                  Oil fields
#> 24       U_EIOU                Elect - Grid                 Crude dist.
#> 25       U_EIOU                Elect - Grid                    NG dist.
#> 26       U_EIOU                Elect - Grid              Oil refineries
#> 27       U_EIOU                Elect - Grid                Power plants
#> 28       U_EIOU                      Diesel              Oil refineries
#> 29       U_EIOU               Crude - Dist.                 Crude dist.
#> 30       U_EIOU              Diesel - Dist.           Gas wells & proc.
#> 31       U_EIOU              Diesel - Dist.                  Oil fields
#> 32       U_EIOU              Diesel - Dist.                 Crude dist.
#> 33       U_EIOU              Diesel - Dist.                    NG dist.
#> 34       U_EIOU              Diesel - Dist.                Petrol dist.
#> 35       U_EIOU                  NG - Wells           Gas wells & proc.
#> 36       U_EIOU              Crude - Fields                  Oil fields
#> 37       U_EIOU              Diesel - Dist.                Diesel dist.
#> 38       U_EIOU              Petrol - Dist.                Petrol dist.
#> 39            Y                  NG - Dist.                 Residential
#> 40            Y              Diesel - Dist.                   Transport
#> 41            Y                Elect - Grid                 Residential
#> 42            Y              Petrol - Dist.                   Transport
#> 43            V           Resources - Crude                       Crude
#> 44            V              Resources - NG                          NG
#> 45            V           Gas wells & proc.                  NG - Wells
#> 46            V                  Oil fields              Crude - Fields
#> 47            V                 Crude dist.               Crude - Dist.
#> 48            V                    NG dist.                  NG - Dist.
#> 49            V              Oil refineries                      Diesel
#> 50            V              Oil refineries                      Petrol
#> 51            V                Power plants                       Elect
#> 52            V                 Elect. grid                Elect - Grid
#> 53            V                Diesel dist.              Diesel - Dist.
#> 54            V                Petrol dist.              Petrol - Dist.
#> 55            V                 Car engines            MD - Car engines
#> 56            V                    Furnaces                         LTH
#> 57            V              Light fixtures                       Light
#> 58            V               Truck engines          MD - Truck engines
#> 59  U_excl_EIOU                       Crude                  Oil fields
#> 60  U_excl_EIOU              Crude - Fields                 Crude dist.
#> 61  U_excl_EIOU               Crude - Dist.              Oil refineries
#> 62  U_excl_EIOU                          NG           Gas wells & proc.
#> 63  U_excl_EIOU                  NG - Wells                    NG dist.
#> 64  U_excl_EIOU                  NG - Dist.                Power plants
#> 65  U_excl_EIOU                  NG - Dist.                    Furnaces
#> 66  U_excl_EIOU                      Diesel                Diesel dist.
#> 67       U_EIOU              Diesel - Dist.           Gas wells & proc.
#> 68       U_EIOU              Diesel - Dist.                  Oil fields
#> 69  U_excl_EIOU              Diesel - Dist.               Truck engines
#> 70  U_excl_EIOU                       Elect                 Elect. grid
#> 71       U_EIOU                Elect - Grid           Gas wells & proc.
#> 72       U_EIOU                Elect - Grid                  Oil fields
#> 73       U_EIOU                Elect - Grid                 Crude dist.
#> 74       U_EIOU                Elect - Grid                    NG dist.
#> 75       U_EIOU                Elect - Grid              Oil refineries
#> 76       U_EIOU                Elect - Grid                Power plants
#> 77  U_excl_EIOU                Elect - Grid              Light fixtures
#> 78  U_excl_EIOU                      Petrol                Petrol dist.
#> 79  U_excl_EIOU              Petrol - Dist.                 Car engines
#> 80       U_EIOU          MD - Truck engines                 Crude dist.
#> 81       U_EIOU          MD - Truck engines                    NG dist.
#> 82       U_EIOU          MD - Truck engines                Diesel dist.
#> 83       U_EIOU          MD - Truck engines                Petrol dist.
#> 84            Y                       Light                 Residential
#> 85            Y                         LTH                   Transport
#> 86            Y            MD - Car engines                 Residential
#> 87            Y          MD - Truck engines                   Transport
#> 88            V           Resources - Crude                       Crude
#> 89            V              Resources - NG                          NG
#> 90            V           Gas wells & proc.                  NG - Wells
#> 91            V                  Oil fields              Crude - Fields
#> 92            V                 Crude dist.               Crude - Dist.
#> 93            V                    NG dist.                  NG - Dist.
#> 94            V              Oil refineries                      Diesel
#> 95            V              Oil refineries                      Petrol
#> 96            V                Power plants                       Elect
#> 97            V                 Elect. grid                Elect - Grid
#> 98            V                Diesel dist.              Diesel - Dist.
#> 99            V                Petrol dist.              Petrol - Dist.
#> 100           V                 Car engines            MD - Car engines
#> 101           V                    Furnaces                         LTH
#> 102           V              Light fixtures                       Light
#> 103           V               Truck engines          MD - Truck engines
#> 104           V                        Cars Passenger [passenger-km/yr]
#> 105           V                       Homes        Space heating [m3-K]
#> 106           V                       Rooms Illumination [lumen-hrs/yr]
#> 107           V                      Trucks     Freight [tonne-km/year]
#> 108 U_excl_EIOU                       Crude                  Oil fields
#> 109 U_excl_EIOU              Crude - Fields                 Crude dist.
#> 110 U_excl_EIOU               Crude - Dist.              Oil refineries
#> 111 U_excl_EIOU                          NG           Gas wells & proc.
#> 112 U_excl_EIOU                  NG - Wells                    NG dist.
#> 113 U_excl_EIOU                  NG - Dist.                Power plants
#> 114 U_excl_EIOU                  NG - Dist.                    Furnaces
#> 115 U_excl_EIOU                      Diesel                Diesel dist.
#> 116      U_EIOU              Diesel - Dist.           Gas wells & proc.
#> 117      U_EIOU              Diesel - Dist.                  Oil fields
#> 118 U_excl_EIOU              Diesel - Dist.               Truck engines
#> 119 U_excl_EIOU                       Elect                 Elect. grid
#> 120      U_EIOU                Elect - Grid           Gas wells & proc.
#> 121      U_EIOU                Elect - Grid                  Oil fields
#> 122      U_EIOU                Elect - Grid                 Crude dist.
#> 123      U_EIOU                Elect - Grid                    NG dist.
#> 124      U_EIOU                Elect - Grid              Oil refineries
#> 125      U_EIOU                Elect - Grid                Power plants
#> 126 U_excl_EIOU                Elect - Grid              Light fixtures
#> 127 U_excl_EIOU                      Petrol                Petrol dist.
#> 128 U_excl_EIOU              Petrol - Dist.                 Car engines
#> 129 U_excl_EIOU                       Light                       Rooms
#> 130 U_excl_EIOU                         LTH                       Homes
#> 131 U_excl_EIOU            MD - Car engines                        Cars
#> 132 U_excl_EIOU          MD - Truck engines                      Trucks
#> 133      U_EIOU     Freight [tonne-km/year]                 Crude dist.
#> 134      U_EIOU     Freight [tonne-km/year]                    NG dist.
#> 135      U_EIOU     Freight [tonne-km/year]                Diesel dist.
#> 136      U_EIOU     Freight [tonne-km/year]                Petrol dist.
#> 137           Y     Freight [tonne-km/year]                   Transport
#> 138           Y Illumination [lumen-hrs/yr]                 Residential
#> 139           Y Passenger [passenger-km/yr]                   Transport
#> 140           Y        Space heating [m3-K]                 Residential
#> 141           V           Resources - Crude                       Crude
#> 142           V              Resources - NG                          NG
#> 143           V           Gas wells & proc.                  NG - Wells
#> 144           V                  Oil fields              Crude - Fields
#> 145           V                 Crude dist.               Crude - Dist.
#> 146           V                    NG dist.                  NG - Dist.
#> 147           V              Oil refineries                      Diesel
#> 148           V              Oil refineries                      Petrol
#> 149           V                Power plants                       Elect
#> 150           V                 Elect. grid                Elect - Grid
#> 151           V                Diesel dist.              Diesel - Dist.
#> 152           V                Petrol dist.              Petrol - Dist.
#> 153           V                 Car engines            MD - Car engines
#> 154           V                    Furnaces                         LTH
#> 155           V              Light fixtures                       Light
#> 156           V               Truck engines          MD - Truck engines
#> 157           V                        Cars Passenger [passenger-km/yr]
#> 158           V                       Homes        Space heating [m3-K]
#> 159           V                       Rooms Illumination [lumen-hrs/yr]
#> 160           V                      Trucks     Freight [tonne-km/year]
#> 161 U_excl_EIOU                       Crude                  Oil fields
#> 162 U_excl_EIOU              Crude - Fields                 Crude dist.
#> 163 U_excl_EIOU               Crude - Dist.              Oil refineries
#> 164 U_excl_EIOU                          NG           Gas wells & proc.
#> 165 U_excl_EIOU                  NG - Wells                    NG dist.
#> 166 U_excl_EIOU                  NG - Dist.                Power plants
#> 167 U_excl_EIOU                  NG - Dist.                    Furnaces
#> 168 U_excl_EIOU                      Diesel                Diesel dist.
#> 169      U_EIOU              Diesel - Dist.           Gas wells & proc.
#> 170      U_EIOU              Diesel - Dist.                  Oil fields
#> 171 U_excl_EIOU              Diesel - Dist.               Truck engines
#> 172 U_excl_EIOU                       Elect                 Elect. grid
#> 173      U_EIOU                Elect - Grid           Gas wells & proc.
#> 174      U_EIOU                Elect - Grid                  Oil fields
#> 175      U_EIOU                Elect - Grid                 Crude dist.
#> 176      U_EIOU                Elect - Grid                    NG dist.
#> 177      U_EIOU                Elect - Grid              Oil refineries
#> 178      U_EIOU                Elect - Grid                Power plants
#> 179 U_excl_EIOU                Elect - Grid              Light fixtures
#> 180 U_excl_EIOU                      Petrol                Petrol dist.
#> 181 U_excl_EIOU              Petrol - Dist.                 Car engines
#> 182 U_excl_EIOU                       Light                       Rooms
#> 183 U_excl_EIOU                         LTH                       Homes
#> 184 U_excl_EIOU            MD - Car engines                        Cars
#> 185 U_excl_EIOU          MD - Truck engines                      Trucks
#> 186      U_EIOU     Freight [tonne-km/year]                 Crude dist.
#> 187      U_EIOU     Freight [tonne-km/year]                    NG dist.
#> 188      U_EIOU     Freight [tonne-km/year]                Diesel dist.
#> 189      U_EIOU     Freight [tonne-km/year]                Petrol dist.
#> 190           Y     Freight [tonne-km/year]                   Transport
#> 191           Y Illumination [lumen-hrs/yr]                 Residential
#> 192           Y Passenger [passenger-km/yr]                   Transport
#> 193           Y        Space heating [m3-K]                 Residential
#>      rowtype  coltype
#> 1   Industry  Product
#> 2   Industry  Product
#> 3   Industry  Product
#> 4   Industry  Product
#> 5   Industry  Product
#> 6   Industry  Product
#> 7   Industry  Product
#> 8   Industry  Product
#> 9   Industry  Product
#> 10  Industry  Product
#> 11  Industry  Product
#> 12  Industry  Product
#> 13   Product Industry
#> 14   Product Industry
#> 15   Product Industry
#> 16   Product Industry
#> 17   Product Industry
#> 18   Product Industry
#> 19   Product Industry
#> 20   Product Industry
#> 21   Product Industry
#> 22   Product Industry
#> 23   Product Industry
#> 24   Product Industry
#> 25   Product Industry
#> 26   Product Industry
#> 27   Product Industry
#> 28   Product Industry
#> 29   Product Industry
#> 30   Product Industry
#> 31   Product Industry
#> 32   Product Industry
#> 33   Product Industry
#> 34   Product Industry
#> 35   Product Industry
#> 36   Product Industry
#> 37   Product Industry
#> 38   Product Industry
#> 39   Product Industry
#> 40   Product Industry
#> 41   Product Industry
#> 42   Product Industry
#> 43  Industry  Product
#> 44  Industry  Product
#> 45  Industry  Product
#> 46  Industry  Product
#> 47  Industry  Product
#> 48  Industry  Product
#> 49  Industry  Product
#> 50  Industry  Product
#> 51  Industry  Product
#> 52  Industry  Product
#> 53  Industry  Product
#> 54  Industry  Product
#> 55  Industry  Product
#> 56  Industry  Product
#> 57  Industry  Product
#> 58  Industry  Product
#> 59   Product Industry
#> 60   Product Industry
#> 61   Product Industry
#> 62   Product Industry
#> 63   Product Industry
#> 64   Product Industry
#> 65   Product Industry
#> 66   Product Industry
#> 67   Product Industry
#> 68   Product Industry
#> 69   Product Industry
#> 70   Product Industry
#> 71   Product Industry
#> 72   Product Industry
#> 73   Product Industry
#> 74   Product Industry
#> 75   Product Industry
#> 76   Product Industry
#> 77   Product Industry
#> 78   Product Industry
#> 79   Product Industry
#> 80   Product Industry
#> 81   Product Industry
#> 82   Product Industry
#> 83   Product Industry
#> 84   Product Industry
#> 85   Product Industry
#> 86   Product Industry
#> 87   Product Industry
#> 88  Industry  Product
#> 89  Industry  Product
#> 90  Industry  Product
#> 91  Industry  Product
#> 92  Industry  Product
#> 93  Industry  Product
#> 94  Industry  Product
#> 95  Industry  Product
#> 96  Industry  Product
#> 97  Industry  Product
#> 98  Industry  Product
#> 99  Industry  Product
#> 100 Industry  Product
#> 101 Industry  Product
#> 102 Industry  Product
#> 103 Industry  Product
#> 104 Industry  Product
#> 105 Industry  Product
#> 106 Industry  Product
#> 107 Industry  Product
#> 108  Product Industry
#> 109  Product Industry
#> 110  Product Industry
#> 111  Product Industry
#> 112  Product Industry
#> 113  Product Industry
#> 114  Product Industry
#> 115  Product Industry
#> 116  Product Industry
#> 117  Product Industry
#> 118  Product Industry
#> 119  Product Industry
#> 120  Product Industry
#> 121  Product Industry
#> 122  Product Industry
#> 123  Product Industry
#> 124  Product Industry
#> 125  Product Industry
#> 126  Product Industry
#> 127  Product Industry
#> 128  Product Industry
#> 129  Product Industry
#> 130  Product Industry
#> 131  Product Industry
#> 132  Product Industry
#> 133  Product Industry
#> 134  Product Industry
#> 135  Product Industry
#> 136  Product Industry
#> 137  Product Industry
#> 138  Product Industry
#> 139  Product Industry
#> 140  Product Industry
#> 141 Industry  Product
#> 142 Industry  Product
#> 143 Industry  Product
#> 144 Industry  Product
#> 145 Industry  Product
#> 146 Industry  Product
#> 147 Industry  Product
#> 148 Industry  Product
#> 149 Industry  Product
#> 150 Industry  Product
#> 151 Industry  Product
#> 152 Industry  Product
#> 153 Industry  Product
#> 154 Industry  Product
#> 155 Industry  Product
#> 156 Industry  Product
#> 157 Industry  Product
#> 158 Industry  Product
#> 159 Industry  Product
#> 160 Industry  Product
#> 161  Product Industry
#> 162  Product Industry
#> 163  Product Industry
#> 164  Product Industry
#> 165  Product Industry
#> 166  Product Industry
#> 167  Product Industry
#> 168  Product Industry
#> 169  Product Industry
#> 170  Product Industry
#> 171  Product Industry
#> 172  Product Industry
#> 173  Product Industry
#> 174  Product Industry
#> 175  Product Industry
#> 176  Product Industry
#> 177  Product Industry
#> 178  Product Industry
#> 179  Product Industry
#> 180  Product Industry
#> 181  Product Industry
#> 182  Product Industry
#> 183  Product Industry
#> 184  Product Industry
#> 185  Product Industry
#> 186  Product Industry
#> 187  Product Industry
#> 188  Product Industry
#> 189  Product Industry
#> 190  Product Industry
#> 191  Product Industry
#> 192  Product Industry
#> 193  Product Industry