Calculate various embodied energy matrices

Calculate various embodied energy matrices

calc_embodied_mats(.iodata = NULL, Y_colname = "Y", q_colname = "q",
  L_ixp_colname = "L_ixp", g_colname = "g", W_colname = "W",
  U_EIOU_colname = "U_EIOU", G_colname = "G", H_colname = "H",
  E_colname = "E", M_p_colname = "M_p", M_s_colname = "M_s",
  F_footprint_p_colname = "F_footprint_p",
  F_effects_p_colname = "F_effects_p",
  F_footprint_s_colname = "F_footprint_s",
  F_effects_s_colname = "F_effects_s")

Arguments

.iodata
a data frame containing matrices that describe the Input-Output structure (using the supply-use table format) of an Energy Conversion Chain. .iodata will likely have been obtained from the calc_io_mats function.

Y_colname
the name of the column in .iodata containing final demand (Y) matrices.

q_colname
the name of the column in .iodata containing final demand (q) vectors.

L_ixp_colname
the name of the column in .iodata containing Industry-by-Product Leontief (L_ixp) matrices.

g_colname
the name of the output column containing g vectors.

W_colname
the name of the output column containing W matrices.

U_EIOU_colname
the name of the output column containing U_EIOU matrices.

G_colname
the name of the output column containing G matrices. G is calculated by L_ixp * y_hat.

H_colname
the name of the output column containing H matrices. H is calculated by L_ixp * Y.

E_colname
the name of the output column containing E matrices. E is calculated by W * g_hat_inv.

M_p_colname
the name of the output column containing M_p matrices. M_p is formed from column sums of positive entries in the various Qx matrices

M_s_colname
the name of the output column containing M_s matrices. M_s is constructed by M_p * q_hat_inv * Y.

F_footprint_p_colname
the name of the output column containing F_footprint_p matrices. F_footprint_p is calculated by M_p (M_p^Ti)_hat_inv.

F_effects_p_colname
the name of the output column containing F_effects_p matrices. F_effects_p is calculated by M_p i_hat_inv M_p.

F_footprint_s_colname
the name of the output column containing F_footprint_s matrices. F_footprint_s is calculated by M_s (M_s^Ti)_hat_inv.

F_effects_s_colname
the name of the output column containing F_effects_s matrices. F_effects_p is calculated by M_s i_hat_inv M_s.

Value

.iodata with columns G_colname, H_colname, E_colname, and Q_colname added