Calculate various input-output matrices

Calculate various input-output matrices

calc_io_mats(.sutdata = NULL, U_colname = "U", V_colname = "V",
  Y_colname = "Y", S_units = "S_units", y_colname = "y",
  q_colname = "q", f_colname = "f", g_colname = "g",
  W_colname = "W", K_colname = "K", Z_colname = "Z",
  C_colname = "C", D_colname = "D", A_colname = "A",
  L_ixp_colname = "L_ixp", L_pxp_colname = "L_pxp")

Arguments

.sutdata
a data frame of supply-use table matrices with matrices arranged in columns.

U_colname
the name of the column in .sutdata containing Use (U) matrices. Default is "U".

V_colname
the name of the column in .sutdata containing Make (V) matrices. Default is "V".

Y_colname
the name of the column in .sutdata containing final demand (Y) matrices. Default is "Y".

S_units
the name of the column in .sutdata containing unit summation (S_units) matrices. Default is "S_units".

y_colname
the name of the output column containing y vectors. Default is "y". y is calculated by rowsums(Y).

q_colname
the name of the output column containing q vectors. Default is "q". q is calculated by rowsums(U) + y.

f_colname
the name of the output column containing f vectors. Default is "f". f is calculated by colsums(U).

g_colname
the name of the output column containing g vectors. Default is "g". g is calculated by rowsums(V).

W_colname
the name of the output column containing W matrices. Default is "W". W is calculated by transpose(V) - U.

K_colname
the name of the output column containing K matrices. Default is "K". K is calculated by U * f_hat_inv.

Z_colname
the name of the output column containing Z matrices. Default is "Z". Z is calculated by U * g_hat_inv.

C_colname
the name of the output column containing C matrices. C is calculated by transpose(V) * g_hat_inv. Default is "C".

D_colname
the name of the output column containing D matrices. Default is "D". D is calculated by V * q_hat_inv.

A_colname
the name of the output column containing A matrices. Default is "A". A is calculated by Z * D.

L_ixp_colname
the name of the output column containing industry-by-product L matrices. Default is "L_ixp". L_ixp is calculated by D * L_pxp.

L_pxp_colname
the name of the output column containing product-by-product L matrices. Default is "L_pxp". L_pxp is calculated by (I - Z*D)^-1.

Value

.sutdata with additional columns: y_colname, q_colname, g_colname, W_colname, Z_colname, D_colname, K_colname, C_colname, A_colname, L_ixp_colname, and L_pxp_colname.