Calculate y, f, g, and q vectors and the W matrix

Note that a necessary condition for calculating the f and g vectors is that the U_bar and V_bar matrices should have only one entry per column and row, respectively, meaning that all products input into a given industry need to be unit homogeneous before we can calculate the f vector and all products of a given industry are measured in the same units before we can calculate the g vector. If the unit homogeneity assumptions above are violated, we will return NA for violating industries in the f and g vectors. The checks for unit homogenity are performed only when an S_units matrix is present.

calc_yqfgW(.sutdata = NULL, U_colname = "U", V_colname = "V",
  Y_colname = "Y", S_units = "S_units", y_colname = "y",
  q_colname = "q", f_colname = "f", g_colname = "g",
  W_colname = "W")

Arguments

.sutdata
a data frame of supply-use table matrices with matrices arranged in columns.

U_colname
the name of the column in .sutdata containing Use (U) matrices.

V_colname
the name of the column in .sutdata containing Make (V) matrices.

Y_colname
the name of the column in .sutdata containing final demand (Y) matrices.

S_units
the name of the column in .sutdata containing S_units matrices.

y_colname
the name of the output column containing y vectors. y is calculated by rowsums_byname(Y).

q_colname
the name of the output column containing q vectors. q is calculated by rowsums_byname(U) + y.

f_colname
the name of the output column containing f vectors. f is calculated by colsums_byname(U).

g_colname
the name of the output column containing g vectors. g is calculated by rowsums_byname(V).

W_colname
the name of the output column containing W matrices. W is calculated by transpose_byname(V) - U.

Value

.sutdata with columns y_colname, q_colname, f_colname, g_colname, and W_colname added