Layout for a qgraph network representation of an energy conversion chain

Industries and products are interleaved from left to right, starting with the first industry, followed by the first product. Industries and products alternate until all products are exhausted. Finally, the last industry (which should be final demand) is placed at the far right. Storage industries are arranged across the top of the network.

ecc_layout(Industries, Products, industry_colname = "Industry",
  product_colname = "Product", stage_colname = "Stage",
  group_colname = "Group", storage_stagename = "Storage", g,
  x_colname = "x", y_colname = "y")

Arguments

Industries
a data frame consisting of columns named industry_colname, stage_colname, and group_colname. Entries in industry_colname must be names of industries. Entries in stage_colname must be names of left-to-right stages in the ECC. Industries (or sectors) in final demand must be included in the Industries data frame. Industries that provide "storage" must also be included in the Industries data frame (such as Stock changes, International aviation and marine bunkers, and Statistical differences), their stage_colname value being storage_stagename. All industries with storage_stagename are laid out along the top of the network. The left-to-right order of stages in the network layout generated by this function is taken from the order of appearance of stages in the stage_colname column. The left-to-right order of storage industries is given by the appearance order of the industries in industry_colname. Entries in group_colname give groups of industries or products to keep together at each stage. The top-to-bottom order of groups is taken from the appearance order in group_colname.

Products
a data frame consisting of same columns as Industries, except that product_colname takes the place of industry_colname. Note that the number of stages in Products must be one less than than number of stages in Industries, not counting industries in stage storage_stagename.

industry_colname
the name of the column in Industries containing names of industries (a string). Default is "Industry".

product_colname
the name of the column in Products containing names of products (a string). Default is "Product".

stage_colname
the name of the columns in Industries and Products containing names of stages (a string). The left-to-right order of stages in the network layout is determined by the top-to-bottom order in which stage names appear in the Industries and Products data frames. Default is "Stage".

group_colname
the name of the column in Industries and Products containing industries and products that should be grouped together vertically at a stage (a string). The top-to-bottom order of groups at a stage is given by the top-to-bottom order of appearance in group_colname. Default is "Group". group_colname is optional in Industries and Products. If group_colname is missing, a column named "group_colname" will be created and filled with "group_colname", thereby creating a single group for all industries or products.

storage_stagename
the name of the stage in stage_colname of Industries that identifies a "storage" industry (a string). Default is "Storage". Typical storage industries are bunkers, stock changes, and statistical differences. This layout puts storage industries across the top of the ECC graph.

g
an optional qgraph object from which the order of node names is extracted, if present. It is assumed that the layout calculated by this function will be applied to g, so the row order of the outgoing layout is set to the row order expected by g. Furthermore, additional error checking is performed: if the names of nodes in Industries and Products don't exactly match the names of nodes that g expects, a data frame of mismatching names is printed and execution is halted. An easy way to obtain a qgraph object for the g argument is to call g <- qgraph(somedata, DoNotPlot = TRUE).

x_colname
the name of the output column containing x coordinates for each node (a string). Default is "x".

y_colname
the name of the output column containing y coordinates for each node (a string). Default is "y".

Value

a matrix with two columns. The first column contains x coordinates for the nodes and is named x_colname. The second column contains y coordinates for the nodes and is named y_colname. Node names are the row names for the outgoing matrix.