Identify industry stages

Names for industries in an ECC may be comprised of a general industry name and the name of specific product, such as "Production - Hydro". Stages of an ECC are most easily identified simply by the general industry name (e.g., all "Production" industries are primary industries), but simple matching of specific industry names with general industry names to identify industry stages will fail ("Production - Hydro" does not match "Production", e.g.). This function assigns ECC stages to specific industry names using a table of general industry names and associated ECC stages. A general industry name matches a specific industry name if the specific industry name starts with or is equal to the general industry name.

identify_industry_stages(GeneralIndustries, SpecificIndustries,
  general_industry_colname = "Industry", stage_colname = "Stage",
  specific_industry_colname = "Industry")

Arguments

GeneralIndustries
a data frame containing at least two columns, general_industry_colname and stage_colname.

SpecificIndustries
a data frame containing at least specific_industry_colname.

general_industry_colname
a string for the name of the column in GeneralIndustries that contains the general names of industries (such as "Production"). Default is "Industry".

stage_colname
a string for the name of the column in GeneralIndustries that contains the names of ECC stages (such as "Primary industry", "Primary-to-final industry", or "Final-to-useful industry"). Default is "Stage".

specific_industry_colname
a string for the name of the column in SpecificIndustries that contains the specific names of industries (such as "Production - Hydro"). Default is "Industry".

Value

a data frame containing specific_industry_colname and stage_colname, wherein the stages associated with general industry names in GeneralIndustries are assigned to the specific industries from SpecificIndustries according to matches between GeneralIndustries$general_industry_colname and SpecificIndustries$specific_industry_colname.

Examples

GeneralIndustries <- data.frame(Industry = c("Production", "Plants", "Cars"),
                                Stage = c("Primary industry", "p-->f industry",
                                          "f-->u industry"), stringsAsFactors = FALSE)
SpecificIndustries <- data.frame(Industry = c("Production - Coal", "Production - Hydro",
                                              "Plants - Electric", "Plants - refinery",
                                              "Cars - Electric", "Cars - Petrol"),
                                              stringsAsFactors = FALSE)
identify_industry_stages(GeneralIndustries = GeneralIndustries,
                         SpecificIndustries = SpecificIndustries)

#>           Industry.x            Stage
#> 1  Production - Coal Primary industry
#> 2 Production - Hydro Primary industry
#> 3  Plants - Electric   p-->f industry
#> 4  Plants - refinery   p-->f industry
#> 5    Cars - Electric   f-->u industry
#> 6      Cars - Petrol   f-->u industry