Assess the effect of new levels of resources

This function calculates the effect of changing the resources available to an energy conversion chain. New versions of U, V, and Y matrices are returned as U_prime, V_prime, and Y_prime.

new_R(.sutdata = NULL, R_prime_colname = "R_prime", U_colname = "U",
  V_colname = "V", Y_colname = "Y", U_prime_colname = "U_prime",
  V_prime_colname = "V_prime", Y_Prime_colname = "Y_prime")

Arguments

.sutdata
a data frame of supply-use table matrices with matrices arranged in columns.

R_prime_colname
the name of an input column in .sutdata containing a new resource matrix for the ECC.

U_colname
the name of a column in .sutdata containing U matrices for the base ECC. Default is "U".

V_colname
the name of a column in .sutdata containing V matrices for the base ECC. Default is "V".

Y_colname
the name of a column in .sutdata containing Y matrices for the base ECC. Default is "Y".

U_prime_colname
the name of the output column that contains new Use (U) matrices. Default is "U_prime".

V_prime_colname
the name of the output column that contains new Make (V) matrices. Default is "V_prime".

Y_Prime_colname

Value

.sutdata with additional columns U_prime_colname, V_prime_colname, and Y_prime_colname.

Details

Note that inputs L_ixp_colname, L_pxp_colname, can be conveniently calculated by the function calc_io_mats.