Reverse an energy conversion chain

Leontief's original input-output analysis involved swimming "upstream" to estimate the economy that would be needed if different final demand were observed. But what if different resources were available? The analysis is the same if resources become final demand (and vice versa) and make becomes use (and vice versa). That is, the analysis is the same if you're dealing with a reversed energy conversion chain (ECC). This function performs that reversal.

reverse(
  .sutmats = NULL,
  R = "R",
  V = "V",
  U = "U",
  Y = "Y",
  suffix = "_rev",
  R_rev = paste0(R, suffix),
  V_rev = paste0(V, suffix),
  U_rev = paste0(U, suffix),
  Y_rev = paste0(Y, suffix)
)

Arguments

.sutmats
the input ECC

R
the R matrix in the ECC to be reversed. (Default is "R".)

V
the V matrix in the ECC to be reversed. (Default is "V".)

U
the U matrix in the ECC to be reversed. (Default is "U".)

Y
the Y matrix in the ECC to be reversed. (Default is "Y".)

suffix
the suffix to be added to matrix names. (Default is "_rev".)

R_rev
the name of the R matrix in the reversed ECC. (Default is "R_rev".)

V_rev
the name of the V matrix in the reversed ECC. (Default is "V_rev".)

U_rev
the name of the U matrix in the reversed ECC. (Default is "U_rev".)

Y_rev
the name of the Y matrix in the reversed ECC. (Default is "Y_rev".)

Value

a reversed version of the ECC described by R, V, U, and Y.

Details

To reverse an ECC, the R, V, U, and Y matrices need to be transposed and swapped: R with Y and V with U. This function performs those operations.

Examples

library(dplyr)
library(Recca)
library(tidyr)
mats <- UKEnergy2000mats %>%
  spread(key = "matrix.name", value = "matrix") %>%
  rename(
    R_plus_V = "V"
  ) %>%
  separate_RV() %>%
  reverse()