Execute expect_known_value interactively or during testing

testthat::expect_known_value is difficult to use, because working directories are different for interactive sessions and during R cmd check. This function abstracts those difficulties by using testthat::is_testing() and other system checks. The assumed directory for saved expected objects is "tests/expectations"

test_against_file(actual_object, expected_file_name, update = FALSE,
  check.attributes = TRUE, expec_folder = file.path("tests",
  "expectations"))

Arguments

actual_object
an object created during a test

expected_file_name
the name of a file previously saved against which actual_object will be tested. This argument must be a string and should probably end in ".rds". Do not include path information. Example: "expected_L.rds".

update
tells whether to update the saved object

check.attributes
tells whether the test should also check attributes during the testing process

expec_folder
a string giving the path to expected values stored as objects on disk. Default is "tests/expectations"

Value

NULL (invisibly). This function is called for its side effect of executing expect_known_value.