Calculates total requirements matrices (L_pxp and L_ixp)

Calculates total requirements matrices (L_pxp and L_ixp)

calc_L(.sutdata = NULL, D_colname = "D", A_colname = "A",
  L_pxp_colname = "L_pxp", L_ixp_colname = "L_ixp")

Arguments

.sutdata
a data frame of supply-use table matrices with matrices arranged in columns.

D_colname
the name of the column in .sutdata containing the D matrix.

A_colname
the name of the column in .sutdata containing the A matrix.

L_pxp_colname
the name of the output column containing the product-by-product L matrix. L_pxp is calculated by (I - Z*D)^-1.

L_ixp_colname
the name of the output column containing the industry-by-product L matrix. L_ixp is calculated by D * L_pxp.

Value

.sutdata with columns L_ixp_colname and L_pxp_colname added