Extend an ECC in PSUT format from energy to exergy

An energy conversion chain can be represented in energy or exergy quantifications of energy. This function moves from an energy quantification to an exergy quantification, given the matrices for the energy quantification and phi (exergy-to-energy ratio) vectors.

Usage

extend_to_exergy(
  .sutmats = NULL,
  clean_up_df = TRUE,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  phi = Recca::psut_cols$phi,
  .exergy_suffix = "_exergy",
  mat_piece = "all",
  phi_piece = "all",
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions_list,
  R_name = Recca::psut_cols$R,
  U_name = Recca::psut_cols$U,
  U_feed_name = Recca::psut_cols$U_feed,
  U_eiou_name = Recca::psut_cols$U_eiou,
  r_eiou_name = Recca::psut_cols$r_eiou,
  V_name = Recca::psut_cols$V,
  Y_name = Recca::psut_cols$Y,
  phi_name = Recca::psut_cols$phi,
  energy_type = Recca::psut_cols$energy_type,
  S_units = Recca::psut_cols$S_units,
  energy = Recca::energy_types$e,
  exergy = Recca::energy_types$x
)

Arguments

.sutmats

An optional data frame of energy conversion chain matrices.

clean_up_df

When .sutmats is a data frame, tells whether to tidyr::pivot_longer() the result, remove no-longer-needed input column phi, and fill the energy_type column with "X" for the exergy versions of the ECC matrices. Default is TRUE.

R, U, U_feed, U_eiou, r_eiou, V, Y, phi

Names of columns in .sutmats or single matrices. See Recca::psut_cols.

.exergy_suffix

The string suffix to be appended to exergy versions of ECC matrices.

mat_piece

The piece of matrix row and column names for R, U, U_feed, U_EIOU, V, and Y matrices which are to be matched against names in the phi vector. Default is "all", meaning that entire names are to be matched.

phi_piece

The piece of names in the phi vector against which row and column names for matrices R, U, U_feed, U_EIOU, V, and Y matrices is to be matched. Default is "all", meaning that entire names are to be matched.

notation

The nomenclature for the row and column labels. Default is RCLabels::bracket_notation.

prepositions

The prepositions to be used row and column notation. Default is RCLabels::prepositions_list.

R_name, U_name, U_feed_name, U_eiou_name, r_eiou_name, V_name, Y_name, phi_name, energy_type, S_units

Names of output matrices

energy, exergy

See Recca::energy_types.

Value

A data frame or list of matrices that represents the exergy version of the ECC.

Details

Internally, this function uses matsindf::apply(), so the ECC matrices can be provided as individual matrices, in a named list, or or in a data frame (in which case the arguments should given the string names of columns in the .sutmats data frame, the default).

The vector phi is considered to be a store of values to be applied to each type of energy carrier. To determine which entry in the phi vector is matched against which energy carrier, mat_piece and phi_piece are consulted. mat_piece and phi_piece can be any of "all", "pref", "suff", "noun", or one of many prepositions given in suffixes

Examples

sutmats <- UKEnergy2000mats %>%
  # Put in wide-by-matrix format.
  tidyr::spread(key = matrix.name, value = matrix) %>%
  # Eliminate services ECCs.
  dplyr::filter(Last.stage %in% c("Final", "Useful")) %>%
  dplyr::mutate(
    phi = RCLabels::make_list(Recca::phi_vec, n = nrow(.), lenx = 1)
  )
extend_to_exergy(sutmats)
#>   Country Year Energy.type Last.stage                  R
#> 1     GBR 2000           E      Final 50000, 0, 0, 43000
#> 2     GBR 2000           E     Useful 50000, 0, 0, 43000
#> 3     GBR 2000           X      Final 53000, 0, 0, 44720
#> 4     GBR 2000           X     Useful 53000, 0, 0, 44720
#>                                          S_units
#> 1             1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 2 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 3             1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 4 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       U
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     0, 500, 47500, 0, 25, 0, 25, 0, 0, 0, 0, 0, 0, 0, 0, 15500, 350, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 50, 0, 25, 43000, 0, 2000, 0, 0, 0, 0, 0, 0, 25, 0, 25, 0, 0, 41000, 0, 0, 50000, 0, 2500, 0, 50, 0, 25, 0, 0, 0, 0, 0, 0, 47000, 0, 5000, 0, 0, 75, 0, 0, 0, 0, 0, 0, 0, 0, 0, 250, 0, 0, 0, 0, 0, 26500, 500, 0, 0, 0, 0, 0, 0, 100, 0, 16000, 0, 0, 0
#> 2 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 26000.0000, 0.0000, 0.0000, 47500.0000, 0.0000, 0.0000, 0.0000, 25.0000, 20.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 15500.0000, 0.0000, 0.0000, 0.0000, 17.9998, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 6400.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 25000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 50.0000, 0.0000, 25.0000, 0.0000, 43000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 6000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 25.0000, 20.0000, 0.0000, 0.0000, 41000.0000, 0.0000, 0.0000, 50000.0000, 0.0000, 0.0000, 0.0000, 50.0000, 0.0000, 25.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 47000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 75.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 26.9997, 0.0000, 0.0000, 0.0000, 26500.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 100.0000, 0.0000, 0.0000, 16000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 15050.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          0.0, 530.0, 50350.0, 0.0, 26.5, 0.0, 25.0, 0.0, NA, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 16430.0, 371.0, 0.0, 0.0, 0.0, NA, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 6400.0, 0.0, 0.0, NA, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 53.0, 0.0, 25.0, 44720.0, NA, 2080.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 26.5, 0.0, 25.0, 0.0, NA, 42640.0, 0.0, 0.0, 53000.0, 0.0, 2650.0, 0.0, 53.0, 0.0, 25.0, 0.0, NA, 0.0, 0.0, 0.0, 0.0, 49820.0, 0.0, 5300.0, 0.0, 0.0, 75.0, 0.0, NA, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 265.0, 0.0, 0.0, 0.0, NA, 0.0, 28090.0, 530.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 100.0, 0.0, NA, 0.0, 0.0, 0.0
#> 4                                                             0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 27560.0000, 0.0000, 0.0000, 50350.0000, 0.0000, 0.0000, 0.0000, 25.0000, 20.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 16430.0000, 0.0000, 0.0000, 0.0000, 17.9998, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 6400.0000, 0.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 53.0000, 0.0000, 25.0000, 0.0000, 44720.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 6000.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 25.0000, 20.0000, 0.0000, NA, 42640.0000, 0.0000, 0.0000, 53000.0000, 0.0000, 0.0000, 0.0000, 53.0000, 0.0000, 25.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 49820.0000, 0.0000, 0.0000, 0.0000, 0.0000, 75.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 26.9997, 0.0000, NA, 0.0000, 28090.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 100.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 15953.0000, 0.0000, 0.0000, 0.0000, 0.0000, NA, 0.0000, 0.0000, 0.0000
#>                                                                                                                                                                                                                                                                                                               U_EIOU
#> 1                                                                                                                 500, 0, 0, 25, 25, 0, 0, 0, 0, 0, 350, 0, 0, 0, 0, 0, 0, 50, 25, 2000, 0, 0, 0, 0, 25, 25, 0, 0, 0, 2500, 0, 50, 25, 0, 0, 0, 0, 5000, 0, 75, 0, 0, 0, 0, 0, 250, 0, 0, 500, 0, 0, 0, 0, 100, 0, 0
#> 2                                                                                                        0.0000, 25.0000, 20.0000, 0.0000, 0.0000, 17.9998, 50.0000, 25.0000, 0.0000, 0.0000, 25.0000, 20.0000, 50.0000, 25.0000, 0.0000, 0.0000, 75.0000, 0.0000, 0.0000, 0.0000, 26.9997, 0.0000, 100.0000, 0.0000
#> 3 530.0, 0.0, 0.0, 26.5, 25.0, 0.0, 0.0, 0.0, 0.0, 0.0, 371.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 53.0, 25.0, 2080.0, 0.0, 0.0, 0.0, 0.0, 26.5, 25.0, 0.0, 0.0, 0.0, 2650.0, 0.0, 53.0, 25.0, 0.0, 0.0, 0.0, 0.0, 5300.0, 0.0, 75.0, 0.0, 0.0, 0.0, 0.0, 0.0, 265.0, 0.0, 0.0, 530.0, 0.0, 0.0, 0.0, 0.0, 100.0, 0.0, 0.0
#> 4                                                                                                        0.0000, 25.0000, 20.0000, 0.0000, 0.0000, 17.9998, 53.0000, 25.0000, 0.0000, 0.0000, 25.0000, 20.0000, 53.0000, 25.0000, 0.0000, 0.0000, 75.0000, 0.0000, 0.0000, 0.0000, 26.9997, 0.0000, 100.0000, 0.0000
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      U_feed
#> 1                                                                                                                                                                                                                                                      0, 0, 47500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 15500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 0, 0, 0, 0, 0, 0, 0, 43000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 41000, 0, 50000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 47000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26500, 0, 0, 0, 0, 0, 0, 16000, 0, 0
#> 2      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26000, 0, 0, 47500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 15500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 25000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 43000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 41000, 0, 0, 50000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 47000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16000, 0, 0, 0, 0, 0, 0, 0, 15050, 0, 0, 0, 0, 0, 0, 0
#> 3                                                                                                                                                                                                                                                 0, 0, 50350, 0, 0, 0, NA, 0, 0, 0, 0, 0, 16430, 0, 0, NA, 0, 0, 0, 0, 0, 0, 6400, 0, NA, 0, 0, 0, 0, 0, 0, 0, 44720, NA, 0, 0, 0, 0, 0, 0, 0, 0, NA, 42640, 0, 53000, 0, 0, 0, 0, 0, NA, 0, 0, 0, 49820, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 28090, 0, 0, 0, 0, 0, 0, NA, 0, 0
#> 4 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 27560, 0, 0, 50350, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 16430, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 44720, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6000, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 42640, 0, 0, 53000, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 49820, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 28090, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 15953, 0, 0, 0, NA, 0, 0, 0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       V
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    47500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 50000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 20500, 0, 0, 0, 15500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 6275, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 41000, 0, 0, 0, 0, 0, 0, 0, 43000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26500, 0
#> 2 0.00, 47000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 47500.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 15500.00, 0.00, 0.00, 0.00, 0.00, 0.00, 15150.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 6400.00, 0.00, 0.00, 0.00, 0.00, 6275.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1200.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 20000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 3000.40, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1799.98, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 41000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 41000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 26500.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 26000.00, 0.00, 0.00
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               50350, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 53000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 21730, 0, 0, 0, 16430, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 6275, 0, 0, 0, 0, 0, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0, 0, 0, 44720, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 28090, 0, 0, 0, 0, 0, 0, 0, 0, 0, 28090, 0
#> 4                                0.00, 49820.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 50350.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 16430.00, 0.00, 0.00, 0.00, 0.00, 0.00, 16059.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 6400.00, 0.00, 0.00, 0.00, 0.00, 6275.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1023.40, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1147.20, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 3000.40, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1799.98, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0.00, 0.00, 0.00, 0.00, 0.00, 42640.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 28090.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 27560.00, 0.00, 0.00
#>                                                             Y
#> 1                       0, 6000, 25000, 0, 14750, 0, 0, 26000
#> 2 1200.00, 0.00, 3000.40, 0.00, 0.00, 20000.00, 0.00, 1714.98
#> 3                         0, 6000, NA, 0, 15635, 0, NA, 27560
#> 4  0.00, 1147.20, 3000.40, 0.00, 1023.40, 0.00, 0.00, 1714.98
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   r_EIOU
#> 1                                                                                                                                                                                                     0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0
#> 2              0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#> 3                                                                                                                                                                                            0, 1, 0, 0, 1, 0, 1, 0, NA, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, NA, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, NA, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, NA, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, NA, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, NA, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, NA, 0, 0, 0
#> 4 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, NA, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, NA, 0, 0, 0