Assess the effect of new levels of resources

This function calculates the effect of changing the resources available to an energy conversion chain. New versions of U, V, and Y matrices are returned as U_prime, V_prime, and Y_prime. This function assumes that each industry's inputs are perfectly substitutable (ps).

new_R_ps(.sutdata = NULL, R_prime_colname = "R_prime",
  U_colname = "U", V_colname = "V", Y_colname = "Y",
  S_units_colname = "S_units", q_colname = "q", C_colname = "C",
  eta_i_colname = "eta_i", maxiter = 100, tol = 0,
  U_prime_colname = "U_prime", V_prime_colname = "V_prime",
  Y_Prime_colname = "Y_prime")

Arguments

.sutdata
a data frame of supply-use table matrices with matrices arranged in columns.

R_prime_colname
the name of an input column in .sutdata containing a new resource matrix for the ECC.

U_colname
the name of a column in .sutdata containing U matrices for the base ECC. Default is "U".

V_colname
the name of a column in .sutdata containing V matrices for the base ECC. Default is "V".

Y_colname
the name of a column in .sutdata containing Y matrices for the base ECC. Default is "Y".

S_units_colname
the name of a column in .sutdata containing S_units matrices for the base ECC. Default is "S_units".

q_colname
the name of a column in .sutdata containing q matrices for the base ECC. Default is "q".

C_colname
the name of a column in .sutdata containing C matrices for the base ECC. Default is "C".

eta_i_colname
the name of a column in .sutdata containing eta_i vectors for the base ECC. Default is "eta_i".

maxiter
the maximum allowable number of iterations when calculating the effects of a new R matrix. Default is 100.

tol
the maximum allowable change in any one entry of the U, V, and Y matrices from one iteration to the next. Default is 0. I.e., when two subsequent iterations must produce the same values, the algorithm has converged.

U_prime_colname
the name of the output column that contains new Use (U) matrices. Default is "U_prime".

V_prime_colname
the name of the output column that contains new Make (V) matrices. Default is "V_prime".

Y_Prime_colname
the name of the output column that contains new Final Demand (Y) matrices. Default is "Y_prime".

Value

.sutdata with additional columns U_prime_colname, V_prime_colname, and Y_prime_colname.

Details

Inputs U_colname, V_colname, Y_colname, S_units_colname, q_colname, and C_colname can be conveniently calculated by the function calc_io_mats; eta_i_colname can be calculated with calc_eta_i.

Each industry must be unit-homogeneous on its inputs. If not, NA is returned as the result for U_prime, V_prime, and Y_prime.

Examples

library(dplyr)
library(matsbyname)
library(tidyr)
doubleR <- UKEnergy2000mats %>%
  spread(key = "matrix.name", value = "matrix") %>%
  # At present, UKEnergy2000mats has V matrices that are the sum of both V and R.
  # Change to use the R matrix.
  rename(
    R_plus_V = V
  ) %>%
  separate_RV() %>%
  # At this point, the matrices are they way we want them.
  # Calculate the input-output matrices which are inputs to the new_R function.
  calc_io_mats() %>%
  # Calculate the efficiency of every industry in the ECC.
  calc_eta_i() %>%
  # Make an R_prime matrix that gives twice the resource inputs to the economy.
  mutate(
    R_prime = elementproduct_byname(2, R)
  ) %>%
  # Now call the new_R function which will calculate
  # updated U, V, and Y matrices (U_prime, V_prime, and Y_prime)
  # given R_prime.
  # Each of the *_prime matrices should be 2x their originals,
  # because R_prime is 2x relative to R.
  # Rows with Last.stage == "services" are NA.
  new_R_ps()