Assess the effect of new levels of resources

This function calculates the effect of changing the resources available to an energy conversion chain. New versions of U, V, and Y matrices are returned as U_prime, V_prime, and Y_prime. This function assumes that each industry's inputs are perfectly substitutable (ps).

new_R_ps(
  .sutmats = NULL,
  R_prime = "R_prime",
  U = "U",
  V = "V",
  Y = "Y",
  S_units = "S_units",
  q = "q",
  C = "C",
  eta_i = "eta_i",
  maxiter = 100,
  tol = 0,
  U_prime = "U_prime",
  V_prime = "V_prime",
  Y_prime = "Y_prime"
)

Arguments

.sutmats
a data frame of supply-use table matrices with matrices arranged in columns.

R_prime
a new resource matrix or name of a column in .sutmats containing same. Default is "R_prime".

U
use (U) matrix or name of the column in .sutmats that contains same. Default is "U".

V
make (V) matrix or name of the column in .sutmatsthat contains same. Default is "V".

Y
final demand (Y) matrix or name of the column in .sutmats that contains same. Default is "Y".

S_units
S_units matrix or name of the column in .sutmats that contains same. Default is "S_units".

q
q vector or name of the column in .sutmats that contains same. Default is "q".

C
a C matrix or name of a column in .sutmats containing same. Default is "C".

eta_i
an eta_i vector or name of a column in .sutmats containing same. Default is "eta_i".

maxiter
the maximum allowable number of iterations when calculating the effects of a new R matrix. Default is 100.

tol
the maximum allowable change in any one entry of the U, V, and Y matrices from one iteration to the next. Default is 0, i.e., when two subsequent iterations produce the same values, the algorithm has converged.

U_prime
name for the U_prime matrix on output. Default is "U_prime".

V_prime
name for the V_prime matrix on output. Default is "V_prime".

Y_prime
name for the Y_prime matrix on output. Default is "Y_prime".

Value

a list or data frame containing U_prime, V_prime, and Y_prime matrices

Details

Inputs U, V, Y, S_units, q, and C can be conveniently calculated by the function calc_io_mats; eta_i can be calculated with calc_eta_i.

Each industry must be unit-homogeneous on its inputs. If not, NA is returned as the result for U_prime, V_prime, and Y_prime.

Examples

library(dplyr)
library(matsbyname)
library(tidyr)
doubleR <- UKEnergy2000mats %>%
  spread(key = "matrix.name", value = "matrix") %>%
  # Calculate the input-output matrices which are inputs to the new_R function.
  calc_io_mats() %>%
  # Calculate the efficiency of every industry in the ECC.
  calc_eta_i() %>%
  # Make an R_prime matrix that gives twice the resource inputs to the economy.
  mutate(
    R_prime = hadamardproduct_byname(2, R)
  ) %>%
  # Now call the new_R function which will calculate
  # updated U, V, and Y matrices (U_prime, V_prime, and Y_prime)
  # given R_prime.
  # Each of the *_prime matrices should be 2x their originals,
  # because R_prime is 2x relative to R.
  # Rows with Last.stage == "services" are NA.
  new_R_ps()
doubleR$U_prime[[1]]

#>                Crude dist. Diesel dist. Elect. grid Gas wells & proc. NG dist.
#> Crude                    0            0           0                 0        0
#> Crude - Dist.         1000            0           0                 0        0
#> Crude - Fields       95000            0           0                 0        0
#> Diesel                   0        31000           0                 0        0
#> Diesel - Dist.          50          700           0               100       50
#> Elect                    0            0       12800                 0        0
#> Elect - Grid            50            0           0                50       50
#> NG                       0            0           0             86000        0
#> NG - Dist.               0            0           0                 0        0
#> NG - Wells               0            0           0              4000    82000
#> Petrol                   0            0           0                 0        0
#> Petrol - Dist.           0            0           0                 0        0
#>                Oil fields Oil refineries Petrol dist. Power plants
#> Crude               1e+05              0            0            0
#> Crude - Dist.       0e+00          94000            0            0
#> Crude - Fields      5e+03              0            0            0
#> Diesel              0e+00          10000            0            0
#> Diesel - Dist.      1e+02              0          500            0
#> Elect               0e+00              0            0            0
#> Elect - Grid        5e+01            150            0          200
#> NG                  0e+00              0            0            0
#> NG - Dist.          0e+00              0            0        32000
#> NG - Wells          0e+00              0            0            0
#> Petrol              0e+00              0        53000            0
#> Petrol - Dist.      0e+00              0         1000            0
#> attr(,"rowtype")
#> [1] "Product"
#> attr(,"coltype")
#> [1] "Industry"

doubleR$V_prime[[1]]

#>                   Crude Crude - Dist. Crude - Fields Diesel Diesel - Dist.
#> Crude dist.           0         95000          0e+00      0              0
#> Diesel dist.          0             0          0e+00      0          31000
#> Elect. grid           0             0          0e+00      0              0
#> Gas wells & proc.     0             0          0e+00      0              0
#> NG dist.              0             0          0e+00      0              0
#> Oil fields            0             0          1e+05      0              0
#> Oil refineries        0             0          0e+00  41000              0
#> Petrol dist.          0             0          0e+00      0              0
#> Power plants          0             0          0e+00      0              0
#>                   Elect Elect - Grid NG NG - Dist. NG - Wells Petrol
#> Crude dist.           0            0  0          0          0      0
#> Diesel dist.          0            0  0          0          0      0
#> Elect. grid           0        12550  0          0          0      0
#> Gas wells & proc.     0            0  0          0      86000      0
#> NG dist.              0            0  0      82000          0      0
#> Oil fields            0            0  0          0          0      0
#> Oil refineries        0            0  0          0          0  53000
#> Petrol dist.          0            0  0          0          0      0
#> Power plants      12800            0  0          0          0      0
#>                   Petrol - Dist.
#> Crude dist.                    0
#> Diesel dist.                   0
#> Elect. grid                    0
#> Gas wells & proc.              0
#> NG dist.                       0
#> Oil fields                     0
#> Oil refineries                 0
#> Petrol dist.               53000
#> Power plants                   0
#> attr(,"rowtype")
#> [1] "Industry"
#> attr(,"coltype")
#> [1] "Product"

doubleR$Y_prime[[1]]

#>                Residential Transport
#> Crude                    0         0
#> Crude - Dist.            0         0
#> Crude - Fields           0         0
#> Diesel                   0         0
#> Diesel - Dist.           0     29500
#> Elect                    0         0
#> Elect - Grid         12000         0
#> NG                       0         0
#> NG - Dist.           50000         0
#> NG - Wells               0         0
#> Petrol                   0         0
#> Petrol - Dist.           0     52000
#> attr(,"rowtype")
#> [1] "Product"
#> attr(,"coltype")
#> [1] "Industry"