Primary aggregate energy

Calculates aggregate primary energy from a data frame of Supply-Use matrices.

primary_aggregates(
  .sutdata,
  p_industries,
  pattern_type = c("exact", "leading", "trailing", "anywhere"),
  R = Recca::psut_cols$R,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  by = c("Total", "Product", "Flow"),
  aggregate_primary = Recca::aggregate_cols$aggregate_primary
)

Arguments

.sutdata
a data frame with columns of matrices from a supply-use analysis.

p_industries
a vector of names of industries to be aggregated as "primary." If .sutdata is a data frame, p_industries should be the name of a column in the data frame. If .sutdata is NULL, p_industries can be a single vector of industry names. These industries in p_industries will appear in rows of the resource (R) and make (V) matrices and columns of the final demand matrix (Y). Entries in Y_p will be subtracted from entries in R_p + V_p to obtain the total primary energy aggregate, where *_p is the primary part of those matrices. The function find_p_industry_names() might be helpful to find primary industry names if they can be identified by prefixes.

pattern_type
One of "exact", "leading", "trailing", or "anywhere" which specifies how matches are made for fd_sectors. If "exact", exact matches specify the sectors to be aggregated. If "leading", sectors are aggregated if any entry in fd_sectors matches the leading part of a final demand sector's name. If "trailing", sectors are aggregated if any entry in fd_sectors matches the trailing part of a final demand sector's name. If "anywhere", sectors are aggregated if any entry in fd_sectors matches any part of a final demand sector's name. Default is "leading" to enable primary industries such as "Resources [of Crude oil]".

R, V, Y
See Recca::psut_cols.

by
One of "Total", "Product", or "Flow" to indicate the desired aggregation:

"Total": aggregation over both Product and Flow (the default)

"Product": aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.)

"Flow": aggregation by type of flow (Production, Imports, Exports, etc.)

aggregate_primary
The name for aggregates of primary energy on output.

Value

A list or data frame containing aggregate primary energy.

Examples

library(matsbyname)
p_industries <- c("Resources - Crude", "Resources - NG")
# Calculate primary total aggregates
res <- UKEnergy2000mats %>%
  tidyr::pivot_wider(names_from = "matrix.name", values_from = "matrix") %>%
  dplyr::mutate(
    p_industries = rep(list(p_industries), times = nrow(.))
  ) %>%
  primary_aggregates(p_industries = "p_industries", by = "Total")

#> Error in primary_aggregates(., p_industries = "p_industries", by = "Total"): unused arguments (p_industries = "p_industries", by = "Total")

res[[Recca::aggregate_cols$aggregate_primary]]

#> Error in eval(expr, envir, enclos): object 'res' not found