Confirm that an SUT-style data frame conserves energy.

Energy balances are confirmed by Product (within tol) for every row in .sutdata.

verify_SUT_energy_balance(.sutdata = NULL, U_colname = "U",
  V_colname = "V", Y_colname = "Y", tol = 1e-06,
  SUT_energy_balance = ".SUT_energy_balance")

Arguments

.sutdata
an SUT-style data frame with columns of matrices, including U, V, and Y columns.

U_colname
the name of the column that contains U matrices. Default is "U".

V_colname
the name of the column that contains V matrices. Default is "V".

Y_colname
the name of the column that contsins Y matrices. Default is "Y".

tol
the maximum amount by which Supply and Consumption can be out of balance. Default is 1e-6.

SUT_energy_balance
the name of the output column. Default is ".SUT_energy_balance".

Value

.sutdata. If energy balance is not observed, an additional column is added showing the row on which energy is not balanced.

Details

If energy is in balance for every row, .sutdata is returned with an additional column, and execution returns to the caller. If energy balance is not observed for one or more of the rows, a warning is emitted, and the additional column (SUT_energy_blance) indicates where the problem occurred, with FALSE showing where energy is not in balance.

This function should be called for its side-effect of testing whether energy is in balance in .sutdata.

Examples

library(dplyr)
library(tidyr)
library(magrittr)
verify_SUT_energy_balance(UKEnergy2000mats %>%
                            filter(Last.stage %in% c("final", "useful")) %>%
                            spread(key = matrix.name, value = matrix),
                          tol = 1e-4)

#>   Country Year Energy.type Last.stage
#> 1      GB 2000      E.ktoe      final
#> 2      GB 2000      E.ktoe     useful
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      r_EIOU
#> 1                                                                                                                                                                                        0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0
#> 2 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
#>                                          S_units
#> 1             1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> 2 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       U
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     0, 500, 47500, 0, 25, 0, 25, 0, 0, 0, 0, 0, 0, 0, 0, 15500, 350, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 50, 0, 25, 43000, 0, 2000, 0, 0, 0, 0, 0, 0, 25, 0, 25, 0, 0, 41000, 0, 0, 50000, 0, 2500, 0, 50, 0, 25, 0, 0, 0, 0, 0, 0, 47000, 0, 5000, 0, 0, 75, 0, 0, 0, 0, 0, 0, 0, 0, 0, 250, 0, 0, 0, 0, 0, 26500, 500, 0, 0, 0, 0, 0, 0, 100, 0, 16000, 0, 0, 0
#> 2 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 26000.0000, 0.0000, 0.0000, 47500.0000, 0.0000, 0.0000, 0.0000, 25.0000, 20.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 15500.0000, 0.0000, 0.0000, 0.0000, 17.9998, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 6400.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 25000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 50.0000, 0.0000, 25.0000, 0.0000, 43000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 6000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 25.0000, 20.0000, 0.0000, 0.0000, 41000.0000, 0.0000, 0.0000, 50000.0000, 0.0000, 0.0000, 0.0000, 50.0000, 0.0000, 25.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 47000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 75.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 26.9997, 0.0000, 0.0000, 0.0000, 26500.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 100.0000, 0.0000, 0.0000, 16000.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 15050.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           V
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  0, 0, 0, 0, 0, 0, 0, 0, 0, 50000, 0, 47500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 50000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 20500, 0, 0, 0, 0, 0, 15500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6400, 0, 0, 0, 0, 6275, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 43000, 0, 0, 0, 0, 41000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 43000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26500, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 26500, 0, 0, 0
#> 2 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 50000.00, 0.00, 0.00, 0.00, 47000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 47500.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 15500.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 15150.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 6400.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 6275.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1200.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 20000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 3000.40, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 1799.98, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 43000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 41000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 41000.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 26500.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 26000.00, 0.00, 0.00, 0.00, 0.00
#>                                                             Y
#> 1                       0, 6000, 25000, 0, 14750, 0, 0, 26000
#> 2 1200.00, 0.00, 3000.40, 0.00, 0.00, 20000.00, 0.00, 1714.98
#>   .SUT_energy_balance
#> 1                TRUE
#> 2                TRUE